% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_analysisQuantifications}
\alias{artms_analysisQuantifications}
\title{Analysis of the Relative Quantification}
\usage{
artms_analysisQuantifications(log2fc_file, modelqc_file, specie,
  enrich = TRUE, output_dir, isFluomics = FALSE, isPtm,
  isBackground = "nobackground", mnbr, l2fc_thres, ipval = "adjpvalue",
  pathogen = "nopathogen")
}
\arguments{
\item{log2fc_file}{(char) MSstats results file location}

\item{modelqc_file}{(char) MSstats modelqc file location}

\item{specie}{(char) Specie (human, mouse)}

\item{enrich}{(logical) Performed enrichment analysis? \code{TRUE} (default) or \code{FALSE}}

\item{output_dir}{(char) Name for the folder to output the results from the function}

\item{isFluomics}{(logical) Is from the fluomics project? \code{TRUE} or \code{FALSE} (default)}

\item{isPtm}{(Is a ptm quantification? \code{noptm} (default), \code{yesptmsites}
(for site specific analysis), \code{yesptmph} (Jeff's script output evidence file)}

\item{isBackground}{(char) background of gene names for enrichment analysis.
\code{nobackground} (default) will use the total number of genes detected.
Alternatively provided the file path name to the background gene list}

\item{mnbr}{(int) minimal number of biological replicates for imputation and filtering}

\item{l2fc_thres}{(int) log2fc cutoff for enrichment analysis}

\item{ipval}{(int) pvalue cutoff for enrichment analysis}

\item{pathogen}{(char) Is there a pathogen in the dataset as well? if it does not,
then use \code{pathogen = nopathogen} (default), \code{tb} (Tuberculosis),
\code{lpn} (Legionella)}
}
\value{
(data.frame) summary of quantifications, including annotations, enrichments, etc
}
\description{
Analysis of the Relative Quantifications obtained by MSstats.
It includes:
\itemize{
\item Annotations
\item Summary files in different format (xls, txt) and shapes (long, wide)
\item Numerous summary plots
\item Enrichment analysis using Gprofiler
\item PCA of protein abundance
\item PCA of quantifications
\item Clustering analysis
}
}
\keyword{analysis,}
\keyword{artms_analysisQuantifications()}
\keyword{quantifications}
