% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{artms_annotationUniprot}
\alias{artms_annotationUniprot}
\title{Uniprot annotations}
\usage{
artms_annotationUniprot(data, columnid, sps)
}
\arguments{
\item{data}{The data.frame to be annotated}

\item{columnid}{The column with the uniprotkb ids}

\item{sps}{The specie name. Check \code{?artms_mapUniprot2entrezGeneName} to
find out more about supported species.}
}
\value{
a data.frame with two new columns: \code{Gene} and \code{Protein.name}
}
\description{
Annotate gene name and symbol based on uniprot ids
}
\keyword{annotation,}
\keyword{artms_annotationUniprot()}
\keyword{uniprot}
