% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_config}
\alias{artms_config}
\title{artMS configuration template}
\format{The configuration (\code{yaml}) file contains the following sections:
\describe{
\item{files}{
\itemize{
\item \code{evidence} : /path/to/the/evidence.txt
\item \code{keys} : /path/to/the/keys.txt
\item \code{contrasts} : /path/to/the/contrast.txt
\item \code{output} : /path/to/the/output/results/results.txt}
}

\item{qc}{
\itemize{
\item \code{enabled} : 1 or 0
}
}

\item{data}{
\itemize{
\item enabled : 1 # 1 = yes; 0 = no
\item fractions:
\itemize{
\item enabled : 0 # 1 for protein fractionation
}
\item silac:
\itemize{
\item enabled : 0 # 1 for SILAC experiments
}
\item filters:
\itemize{
\item enabled : 1
}
\item contaminants : 1
\item protein_groups : remove #remove, keep
\item modifications : ab # PH, UB, AB, APMS
\item sample_plots : 1 # correlation plots }
}

\item{msstats}{
\itemize{
\item enabled : 1
\item msstats_input : # blank if not previous msstats input file is available
\item profilePlots : none # before, after, before-after, none
\item normalization_method : equalizeMedians # globalStandards (include a reference protein(s) ), equalizeMedians, quantile, 0
\item normalization_reference :  #should be a value in the Protein column
\item summaryMethod : TMP # "TMP"(default) means Tukey's median polish, which is robust estimation method. "linear" uses linear mixed model. "logOfSum" conducts log2 (sum of intensities) per run.
\item censoredInt : NA  # Missing values are censored or at random. 'NA' (default) assumes that all 'NA's in 'Intensity' column are censored. '0' uses zero intensities as censored intensity. In this case, NA intensities are missing at random. The output from Skyline should use '0'. Null assumes that all NA intensites are randomly missing.
\item cutoffCensored : minFeature  # Cutoff value for censoring. only with censoredInt='NA' or '0'. Default is 'minFeature', which uses minimum value for each feature.'minFeatureNRun' uses the smallest between minimum value of corresponding feature and minimum value of corresponding run. 'minRun' uses minumum value for each run.
\item MBimpute : 1 # only for summaryMethod="TMP" and censoredInt='NA' or '0'. TRUE (default) imputes 'NA' or '0' (depending on censoredInt option) by Accelated failure model. FALSE uses the values assigned by cutoffCensored.
\item feature_subset: all # all|highQuality  : highQuality seems to be buggy right now
}
}

\item{output_extras}{
\itemize{
\item enabled : 0
\item msstats_output : # keep empty if mstats is enabled. Or provide the mstats 'results.txt' file
\item annotate : 0 # 1|0 whether to annotate the proteins in the results or not
\item species : HUMAN  # can use multiple species, but separate with a "-" eg. HUMAN-MOUSE-HIV-...
\item annotation_dir : /path/to/the/files
\item comparisons : all # or any grep expression that returns a subset of the contrasts file
\item LFC : -2 2
\item FDR : 0.05
\item heatmap : 1
\item heatmap_cluster_cols : 0
\item heatmap_display : log2FC #or pvalue
\item volcano : 1}
}
}}
\usage{
artms_config
}
\description{
The configuration file in \code{yaml} format contains
the configuration details required to run \code{artms_quantification()}, which
includes quality control functions
}
\keyword{datasets}
