% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_corum_mito_database}
\alias{artms_data_corum_mito_database}
\title{CORUM Protein Complexes database use for complex enrichment analysis}
\format{Tab delimited file.
\describe{
To find out more about the format and columns available at CORUM,
please visit this
\href{http://mips.helmholtz-muenchen.de/corum/}{link}
}}
\usage{
artms_data_corum_mito_database
}
\description{
The list of protein complexes has been enriched with
mitochondria proteins from mouse, as described in this paper:

2018 - Ruchi Masand, Esther Paulo, Dongmei Wu , Yangmeng Wang,
Danielle L. Swaney, David Jimenez-Morales, Nevan J. Krogan, and Biao Wang
Proteome Imbalance of Mitochondrial Electron Transport Chain in Brown
Adipocytes Leads to Metabolic Benefits.
Cell Metab. 2018 Mar 06; 27(3):616-629.e4
}
\details{
LAST CORUM DOWNLOAD DATE: 2017-08-01
}
\keyword{datasets}
