% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_pathogen_LPN}
\alias{artms_data_pathogen_LPN}
\title{LPN PATHOGEN: Legionella pneumophila subsp. pneumophila
(strain Philadelphia 1 / ATCC 33152 / DSM 7513) UNIPROT IDS}
\format{A data.frame of Entry IDs}
\usage{
artms_data_pathogen_LPN
}
\description{
LPN PATHOGEN: Legionella pneumophila subsp. pneumophila
(strain Philadelphia 1 / ATCC 33152 / DSM 7513) UNIPROT IDS
}
\keyword{datasets}
