% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_pathogen_TB}
\alias{artms_data_pathogen_TB}
\title{TB PATHOGEN: Mycobacterium tuberculosis
(strain ATCC 35801 / TMC 107 / Erdman) UNIPROTS IDS}
\format{
A data.frame of Entry IDs
}
\usage{
artms_data_pathogen_TB
}
\description{
TB PATHOGEN: Mycobacterium tuberculosis
(strain ATCC 35801 / TMC 107 / Erdman) UNIPROTS IDS
}
\keyword{datasets}
