% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_config}
\alias{artms_data_ph_config}
\title{artMS configuration for the available PH dataset}
\format{
The configuration (\code{yaml}) file contains the following sections:
\describe{
\item{files}{
\itemize{
\item \code{evidence} : artms_data_ph_evidence
\item \code{keys} : artms_data_ph_keys
\item \code{contrasts} : artms_data_ph_contrast
\item \code{summary} :
\item \code{output} : "results.txt"
}
}

\item{qc}{
\itemize{
\item basic: 0
\item extended: 0
\item extendedSummary: 0 =
}
}

\item{data}{
\itemize{
\item enabled : 1
\item fractions:
\itemize{
\item enabled : 0
}
\item silac:
\itemize{
\item enabled : 0
}
\item filters:
\itemize{
\item enabled : 1
}
\item contaminants : 1
\item protein_groups : remove
\item modifications : PH
\item sample_plots : 1
}
}

\item{msstats}{
\itemize{
\item enabled : 1
\item msstats_input : # blank if not previous msstats input file is available
\item profilePlots : none # before, after, before-after, none
\item normalization_method : equalizeMedians
\item normalization_reference :  #should be a value in the Protein column
\item summaryMethod : TMP
\item censoredInt : NA
\item cutoffCensored : minFeature
\item MBimpute : 1
\item feature_subset: all
}
}

\item{output_extras}{
\itemize{
\item output_extras :
\itemize{
\item enabled : 1
}
\item annotate :
\itemize{
\item enabled: 1
}
\item species : HUMAN
\item plots:
\itemize{
\item volcano: 1
\item heatmap: 1
\item LFC : -1 1
\item FDR : 0.05
\item heatmap_cluster_cols : 0
\item heatmap_display : log2FC}
}
}
}
}
\usage{
artms_data_ph_config
}
\description{
The configuration file with default options to run the
available PH dataset with `artmsQuantification()``
}
\keyword{datasets}
