% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_contrast}
\alias{artms_data_ph_contrast}
\title{Contrast example for the PH dataset}
\format{
list with one comparison: "Cal33-HSC6"
}
\usage{
artms_data_ph_contrast
}
\description{
Contrast file with the relative quantification to be performed
for the two conditions available in the example dataset: "Cal33-HSC6".
See vignette for more details on how to prepare the contrast file.
}
\keyword{datasets}
