% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_msstats_modelqc}
\alias{artms_data_ph_msstats_modelqc}
\title{MSstats modelQC example}
\format{A data frame resulting from running the lastest version of
\code{MSstats::groupComparison} function required as input for
artmsAnalysisQuantifications()}
\usage{
artms_data_ph_msstats_modelqc
}
\description{
Normalized data obtained from the \code{artmsQuantification()} step
of the PH dataset (global analysis)
}
\keyword{datasets}
