% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_msstats_results}
\alias{artms_data_ph_msstats_results}
\title{MSstats results example}
\format{
A data frame resulting from running the lastest version of MSstats
}
\usage{
artms_data_ph_msstats_results
}
\description{
Relative quantification results obtained running MSstats
on the available PH datasets (global analysis).
Changes in protein phosphorylation were quantified between two conditions
(check \code{artms_data_ph_contrast})
}
\keyword{datasets}
