% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_dataPlots}
\alias{artms_dataPlots}
\title{Protein abundance dot plots}
\usage{
artms_dataPlots(input_file, output_file)
}
\arguments{
\item{input_file}{The \code{-normalized.txt} output file from MSstats}

\item{output_file}{Wished output file name (add the \code{.pdf} extension)}
}
\value{
A pdf file with each individual protein abundance plot for each
conditions
}
\description{
Protein abundance dot plots for each unique uniprot id
}
\keyword{abundance,}
\keyword{artms_dataPlots()}
\keyword{dotplots,}
\keyword{plot}
