% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichForComplexes}
\alias{artms_enrichForComplexes}
\title{Enrich for Protein Complexes using CORUM}
\usage{
artms_enrichForComplexes(df, backgroundNumber)
}
\arguments{
\item{df}{Data.frame with columns: \code{Protein} and \code{Conditions}}

\item{backgroundNumber}{Background number of genes}
}
\value{
A text delimited data.frame with protein complex enrichment results
}
\description{
Enrich for Protein Complexes using CORUM
}
\keyword{artms_enrichForComplexes()}
\keyword{complexes}
\keyword{enrichment,}
\keyword{protein,}
