% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichLog2fc}
\alias{artms_enrichLog2fc}
\title{Enrichment of changes in protein abundance or PTMs}
\usage{
artms_enrichLog2fc(data, fileout, specie, background)
}
\arguments{
\item{data}{(data.frame) with \code{Gene} and \code{Comparison} columns}

\item{fileout}{(char) Name for the output file}

\item{specie}{(char) Specie, only supported "human" or "mouse"}

\item{background}{(vector) Background genes for the enrichment analysis.}
}
\value{
(data.frame) Results from the enrichment analysis using Gprofiler
}
\description{
Enrichment analysis of the selected proteins
}
\examples{
\donttest{
artms_enrichLog2fc(data = df, 
                   fileout = "results-enrich.txt", 
                   specie = "human", 
                   background = df$Gene)
}
}
\keyword{enrichment}
