% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichProfiler}
\alias{artms_enrichProfiler}
\title{Enrichment analysis using GprofileR}
\usage{
artms_enrichProfiler(x, categorySource = c("GO"), specie,
  background = NA)
}
\arguments{
\item{x}{List of protein ids. \code{x} can be anything: either a list of ids,
or you could also send a data.frame and it will find the columns
with the IDs. Is not cool?}

\item{categorySource}{Resources providing the terms on which the enrichment
will be performed. The supported resources by gprofiler are:
\itemize{
\item \emph{GO (GO:BP, GO:MF, GO:CC)}: Gene Ontology (see more below)
\item \emph{KEGG}: Biological pathways
\item \emph{REAC}: Biological pathways (Reactome)
\item \emph{TF}: Regulatory motifs in DNA (TRANSFAC TFBS)
\item \emph{MI}: Regulatory motifs in DNA (miRBase microRNAs)
\item \emph{CORUM}: protein complexes database
\item \emph{HP}: Human Phenotype Ontology
\item \emph{HPA}: Protein databases (Human Protein Atlas)
\item \emph{OMIM}: Online Mendelian Inheritance in Man annotations:
\item \emph{BIOGRID}: BioGRID protein-protein interactions
The type of annotations for Gene Ontology:
\item Inferred from experiment \link[=*IDA, IPI, IMP, IGI, IEP*]{IDA, IPI, IMP, IGI, IEP}
\item Direct assay \link[=*IDA*]{IDA} / Mutant phenotype \link[=*IMP*]{IMP}
\item Genetic interaction \link[=*IGI*]{IGI} / Physical interaction \link[=*IPI*]{IPI}
\item Traceable author \link[=*TAS*]{TAS} / Non-traceable author \link[=*NAS*]{NAS} /
Inferred by curator \link[=*IC*]{IC}
\item Expression pattern \link[=*IEP*]{IEP} / Sequence or structural similarity \link[=*ISS*]{ISS}
/ Genomic context \link[=*IGC*]{IGC}
\item Biological aspect of ancestor \link[=*IBA*]{IBA} / Rapid divergence \link[=*IRD*]{IRD}
\item Reviewed computational analysis \link[=*RCA*]{RCA} / Electronic annotation \link[=*IEA*]{IEA}
\item No biological data \link[=*ND*]{ND} / Not annotated or not in background \link[=*NA*]{NA}
}}

\item{specie}{Specie code: Organism names are constructed by concatenating
the first letter of the name and the family name.
Example: human - ’hsapiens’, mouse - ’mmusculus’.}

\item{background}{Vector gene list to use as background for the enrichment
analysis. Default: \code{NA}}
}
\value{
The enrichment results as provided by gprofiler
}
\description{
This function simplifies the enrichment analysis performed by
the excellent tool GprofileR.
}
\details{
This function uses the following \code{gprofiler} arguments as default:
\itemize{
\item ordered_query = F
\item significant = T
\item exclude_iea = T
\item underrep = F
\item evcodes = F
\item region_query = F
\item max_p_value = 0.05
\item min_set_size = 0
\item max_set_size = 0
\item min_isect_size = 0
\item correction_method = "analytical" #Options: "gSCS", "fdr", "bonferroni"
\item hier_filtering = "none"
\item domain_size = "known" # annotated or known
\item numeric_ns = ""
\item png_fn = NULL
\item include_graph = T
}
}
\keyword{artms_enrichProfiler()}
\keyword{enrichment}
