% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceQC.R
\name{artms_evidenceQC.R}
\alias{artms_evidenceQC.R}
\title{Quality Control analysis of the MaxQuant evidence file}
\usage{
artms_evidenceQC.R(evidence_file, keys_file, prot_exp, output_dir, fractions)
}
\arguments{
\item{evidence_file}{The evidence file}

\item{keys_file}{The keys file with the experimental design}

\item{prot_exp}{Proteomics experiment. 4 options available:
\itemize{
\item \code{apms}: affinity purification mass spectrometry
\item \code{ab}: protein abundance
\item \code{ph}: protein phosphorylation
\item \code{ub}: protein ubiquitination (aka ubiquitylation)
}}

\item{output_dir}{output directory to save results in}

\item{fractions}{Is a fractionated experiment?
\itemize{
\item 1 yes
\item 0 no (default)
}}
}
\value{
Quality control files and plots
}
\description{
Quality Control analysis of the MaxQuant evidence file
}
\keyword{QC,}
\keyword{artms_evidenceQC()}
\keyword{control,}
\keyword{evidence}
\keyword{quality,}
