% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToMISTformat.R
\name{artms_evidenceToMISTformat}
\alias{artms_evidenceToMISTformat}
\title{Convert MaxQuant Evidence file into a Protein Prospector like format
to run through the MIST pipeline using spectral counts.}
\usage{
artms_evidenceToMISTformat(input_file, quant_variable, keys_file,
  output_file, species, uniprot_dir)
}
\arguments{
\item{input_file}{(char) MaxQuant evidence file and location}

\item{quant_variable}{(char) Select the quantitative variable.
Two options available:
\itemize{
\item \code{msint}: MS Intensity
\item \code{msspc}: MS.MS.count (Spectral Counts)
}}

\item{keys_file}{(char) Keys file with the experimental details}

\item{output_file}{(char) Output file name}

\item{species}{(char) Specie name. If several, used a \code{dash} symbol
to separate them}

\item{uniprot_dir}{Directory with the uniprot files with the mapping
information. Default '~/Box Sync/db/mist/'}
}
\value{
MIST compatible input files (reformatted evidence and keys)
}
\description{
Converts MaxQuant evidence file into a file format compatible
with the MiST pipeline using \code{MS/MS.Count}. Note that this is the
MiST \emph{data} file, and that an additional \emph{keys} file will have to be
constructed before running MiST. Multiple species can be searched at once,
simply separate them by a "-". (eg. \code{HUMAN-MOUSE})
}
\examples{
\donttest{
artms_evidenceToMISTformat(input_file = "a549-PB1-evidence.txt",
                           keys_file = "a549-PB1-keys.txt",
                           quant_variable = "msint",
                           output_file = "a549-PB1-mist.txt",
                           species = "HUMAN-FLUOMICS",
                           uniprot_dir = "~/Box Sync/db/mist/")
}
}
\keyword{apms}
\keyword{evidence,}
\keyword{keys,}
\keyword{mist,}
