% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToMISTformat.R
\name{artms_evidenceToMISTINTformat}
\alias{artms_evidenceToMISTINTformat}
\title{Convert MaxQuant file into a Protein Prospector like format to run
through the Mist pipeline using MS1 Intensity values.}
\usage{
artms_evidenceToMISTINTformat(input_file, keys_file, output_file, species,
  uniprot_dir)
}
\arguments{
\item{input_file}{MaxQuant evidence file and location}

\item{keys_file}{Keys file with the experimental details}

\item{output_file}{Output file name}

\item{species}{Specie name. If several, used a \code{dash} symbol to separate them}

\item{uniprot_dir}{Directory with the uniprot files with the mapping
information. Default '~/Box Sync/db/mist/'}
}
\value{
A MIST compatible input file
}
\description{
Converts MaxQuant evidence file into a file format compatible
with the MiST pipeline using \code{Intensity}. Note that this is the
MiST \emph{data} file, and that an additional \emph{keys} file will have to be
constructed before running MiST. Multiple species can be searched at once,
simply separate them by a "-". (eg. \code{HUMAN-MOUSE})
}
\keyword{apms}
\keyword{artms_evidenceToMISTINTformat()}
\keyword{evidence,}
\keyword{keys,}
\keyword{mist,}
