% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToSAINTqFormat.R
\name{artms_evidenceToSAINTqFormat}
\alias{artms_evidenceToSAINTqFormat}
\title{MaxQuant evidence file to SAINTq format}
\usage{
artms_evidenceToSAINTqFormat(evidence_file, keys_file, output_dir,
  sc_option = "all", fractions = FALSE, quant_variable = "msint")
}
\arguments{
\item{evidence_file}{(char) Maxquant evidence file}

\item{keys_file}{(char) Keys file with a SAINT column specifying
test (\code{T}) and control (\code{C}) conditions}

\item{output_dir}{(char) New directory to create ans save files}

\item{sc_option}{(char). Filter peptides with spectral counts only.
Two options:
\itemize{
\item \code{msspc}: use only peptides with spectral_counts
\item \code{all} (default): all peptides detected (including the one resulting from the
MaxQuant 'Match between run' algorithm)
}}

\item{fractions}{(logical) \code{TRUE} for 2D proteomics (fractions).
Default: \code{FALSE}}

\item{quant_variable}{(char) Select the quantitative variable.
Two options available:
\itemize{
\item \code{msint}: MS Intensity
\item \code{msspc}: MS.MS.count (Spectral Counts)
}}
}
\value{
The input files requires to run SAINTq
}
\description{
Converts the MaxQuant evidence file to the 3 required files
by SAINTexpress. One can choose to either use the \code{spectral counts}
(use \code{msspc}) or the \code{intensities} (use \code{msint}) for the analysis.
}
\details{
After running the script, the new specified folder should contain
the folling files:
\itemize{
\item saintq-config-peptides
\item saintq-config-proteins
\item saintq_input_peptides.txt
\item saintq_input_proteins.txt
}

Then \code{cd} into the new folder and run either of the following two options
(assuming that \code{saintq} is installed in your linux/unix/mac os x system):

\code{> saintq config-saintq-peptides}

or

\code{> saintq config-saintq-proteins}
}
\examples{
\donttest{
artms_evidenceToSAINTqFormat(evidence_file = "a549-PB1-evidence.txt", 
                             keys_file = "a549-PB1-keys.txt", 
                             output_dir = "saintq_folder")
}
}
\keyword{APMS}
\keyword{SAINT,}
\keyword{SAINTq,}
