% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_filterEvidenceContaminants}
\alias{artms_filterEvidenceContaminants}
\title{Remove contaminants and empty proteins from the MaxQuant evidence file}
\usage{
artms_filterEvidenceContaminants(data)
}
\arguments{
\item{data}{(data.frame) of the Evidence file}
}
\value{
(data.frame) without REV__ and CON__ Protein ids
}
\description{
Remove contaminants and erronously identified 'reverse'
sequences by MaxQuant, in addition to empty protein ids
}
\examples{
ef <- artms_filterEvidenceContaminants(data = artms_data_ph_evidence)
}
\keyword{cleanup,}
\keyword{contaminants}
