% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOnNumUnique.R
\name{artms_filterNumUniqPep}
\alias{artms_filterNumUniqPep}
\title{Filter MaxQuant data based on Number of Unique Peptides}
\usage{
artms_filterNumUniqPep(evidence_file, keys_file, prot_groups_file, min_pep,
  verbose = TRUE, return.results = FALSE)
}
\arguments{
\item{evidence_file}{(char) The filepath to the MaxQuant searched data
(evidence) file (txt tab delimited file).}

\item{keys_file}{(char) The filepath to the keys file. The condition names
must match the Condition names provided to the proteinGroups file replicates.}

\item{prot_groups_file}{The filepath to the MaxQuant searched
\code{proteinGroups.txt} (txt tab delimited file).}

\item{min_pep}{(int) The minimum number of peptides that need
to be found with at least one bait-prey protein pair to consider it a
real hit. Anything not meeting this criteria is removed before scoring.}

\item{verbose}{(Boolean) Whether or not to display status reports.
Default = TRUE}

\item{return.results}{(Boolean) Whether or not to return the results to
the R environment upon completion. This is useful if this is being used
in an R pipeline and you want to feed the results directly into the next
stage of analysis via an R environment/terminal.
Regardless, the results will be written to file. Default = FALSE}
}
\description{
Remove all condition-protein instance where the number of
unique peptides is below the \code{min_pep} threshold.
The results are written out to a file that is named similar to the evidence
file, but also contains the filtering cutoff used.
}
\keyword{MSStats}
\keyword{MaxQuant,}
\keyword{artms_filterNumUniqPep()}
\keyword{evidence,}
