% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_generatePhSiteExtended}
\alias{artms_generatePhSiteExtended}
\title{Generate ph-site specific detailed file}
\usage{
artms_generatePhSiteExtended(df, pathogen, specie)
}
\arguments{
\item{df}{(data.frame) of log2fc and imputed values}

\item{pathogen}{(char) Is there a pathogen in the dataset as well? Available
pathogens are \code{tb} (Tuberculosis), \code{lpn} (Legionella). If it is not,
then use \code{nopathogen} (default).}

\item{specie}{(char) Main organism (supported for now: \code{human} or \code{mouse})}
}
\value{
(data.frame) extended version of the ph-site
}
\description{
Generate extended detailed ph-site file, where every line is a
ph site instead of a peptide. Therefore, if one peptide has multiple ph sites
it will be breaking down in each of the sites. This file will help generate
input files for tools as \href{http://phosfate.com/}{Phosfate} or
\href{https://github.com/jdrudolph/photon}{PHOTON}
}
\keyword{artms_generatePhSiteExtended()}
\keyword{external,}
\keyword{phosfate}
\keyword{tools,}
