% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_imputeMissingValues}
\alias{artms_imputeMissingValues}
\title{Imputing missing values}
\usage{
artms_imputeMissingValues(dflog2fcinfinites, dfmq)
}
\arguments{
\item{dflog2fcinfinites}{data.frame of proteins with only infinite
values from the msstats results file}

\item{dfmq}{Abundance data, which will be used to know the details of
reproducibility}
}
\value{
Imputed missing values
}
\description{
When a value is completely missed in one of the conditions,
the \code{log2fc = Inf / -Inf}. This function imputes those values, i.e.,
will assign 'artificial' values.
The imputation method works as follow. The assumption is that those
proteins are likely present as well in those conditions where are found
missed, but due to the small sampling (usually 2 or 3 biological replicas)
and other proteomics related issues, those proteins didn't make it through
the level of detection.
Therefore, a small intensity (sampled from the bottom 1%) will be assigned
to the protein/site in the missing condition, and the new log2fc is
re-calculated out of the MSstats box. Two issues are addressed as follows:
\enumerate{
\item If a protein has been consistently identified in one of the conditions,
it will stay
\item But if the intensity value in those conditions was too low,
then the log2fc will be also low
}
}
\keyword{artms_imputeMissingValues()}
\keyword{imputation,}
\keyword{log2fc,}
\keyword{missing}
\keyword{quantifications,}
\keyword{values}
