% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfFile.R
\name{artms_isEvidenceNewVersion}
\alias{artms_isEvidenceNewVersion}
\title{Check if a given evidencee file was generated by a new version of
MaxQuant.}
\usage{
artms_isEvidenceNewVersion(evidence_file)
}
\arguments{
\item{evidence_file}{the evidence file name}
}
\description{
MaxQuant introduced changes in the column names and number
of columns for the evidence file. This function check whether the evidence
it comes from the latest version of the evidence file.
}
\examples{
\donttest{
artms_isEvidenceNewVersion(evidence_file = "/path/to/evidence.txt")
}
}
\keyword{check}
\keyword{evidence,}
\keyword{file,}
\keyword{input,}
\keyword{version}
