% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_loadModelQCstrict}
\alias{artms_loadModelQCstrict}
\title{Load limited columns from abundance (modelqc) annotated}
\usage{
artms_loadModelQCstrict(df_input, specie, isPTM)
}
\arguments{
\item{df_input}{data.frame with the raw abundance data (modelqc)}

\item{specie}{Specie name for annotation purposes}

\item{isPTM}{Specify whether is a PTM dataset: noptmsites, yesptmsites, yesptmph}
}
\value{
annotated data.frame of abundance data
}
\description{
Load limited columns from abundance (modelqc) annotated
}
\keyword{abundance,}
\keyword{annotated}
\keyword{artms_loadModelQCstrict()}
