% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_main.R, R/main.R
\name{artms_main}
\alias{artms_main}
\alias{artms_main}
\title{Main Function}
\usage{
artms_main(yaml_config_file)

artms_main(yaml_config_file)
}
\arguments{
\item{yaml_config_file}{The yaml file name and location}

\item{yaml_config_file}{The yaml file name and location}
}
\value{
All the selected options

All the selected options
}
\description{
Main function running all the selected options

Main function running all the selected options
}
\keyword{artms_main()}
\keyword{driver,}
\keyword{function}
\keyword{main,}
