% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{artms_mapUniprot2entrezGeneName}
\alias{artms_mapUniprot2entrezGeneName}
\title{Map GENE SYMBOL, NAME, AND ENTREZID to a vector of Uniprot IDS}
\usage{
artms_mapUniprot2entrezGeneName(uniprotkb, specie)
}
\arguments{
\item{uniprotkb}{(vector) Vector of UniprotKB IDs}

\item{specie}{(char) The specie name. Currently supporting:
\itemize{
\item ANOPHELES
\item ARABIDOPSIS
\item BOVINE
\item WORM
\item CANINE
\item FLY
\item ZEBRAFISH
\item ECOLI_STRAIN_K12
\item ECOLI_STRAIN_SAKAI
\item CHICKEN
\item HUMAN
\item MOUSE
\item RHESUS
\item MALARIA
\item CHIMP
\item RAT
\item PIG
\item XENOPUS
}}
}
\value{
(data.frame) with EntrezID and GENENAMES mapped on UniprotKB ids
}
\description{
Map GENE SYMBOL, NAME, AND ENTREZID to a vector of Uniprot IDS
}
\examples{
uniprots_anno <- artms_mapUniprot2entrezGeneName(
                       uniprotkb = unique(artms_data_ph_evidence$Proteins),
                       specie = 'human')
}
\keyword{annotation,}
\keyword{ids}
