% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_mergeAbNbr}
\alias{artms_mergeAbNbr}
\title{Merge abundance and number of biological replicates per condition}
\usage{
artms_mergeAbNbr(df_input, repro, specie)
}
\arguments{
\item{df_input}{Abundance input file}

\item{repro}{Reproducibility data.frame}

\item{specie}{Specie for annotation purposes}
}
\value{
data.frame of abundance merged with reproducibility info
}
\description{
Merge abundance and number of biological replicates
per condition
}
\keyword{abundance,}
\keyword{artms_mergeAbNbr()}
\keyword{merging}
\keyword{reproducibility,}
