% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_mergeChangesNbr}
\alias{artms_mergeChangesNbr}
\title{Merge changes (log2fc) and number of biological replicates per
condition}
\usage{
artms_mergeChangesNbr(df_input, repro, specie)
}
\arguments{
\item{df_input}{Changes data.frame}

\item{repro}{Reproducibility data.frame}
}
\value{
specie Specie for annotation purposes
}
\description{
Merge changes, i.e., MSstats results file of quantified changes,
with the number of biological replicates per condition
}
\keyword{artms_mergeChangesNbr()}
\keyword{changes,}
\keyword{log2fc,}
\keyword{merging}
\keyword{reproducibility,}
