% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_mergeEvidenceAndKeys}
\alias{artms_mergeEvidenceAndKeys}
\title{Merge evidence and keys files}
\usage{
artms_mergeEvidenceAndKeys(data, keys, by = c("RawFile"))
}
\arguments{
\item{data}{(data.frame or char) The evidence data, either as adata.frame or
the file name (and path)}

\item{keys}{The keys data, either as a data.frame or file name (and path)}

\item{by}{(vector) specifying the columns use to merge the evidence and keys.
Default: \code{by=c('RawFile')}}
}
\value{
(data.frame) with the evidence and keys merged
}
\description{
Merge the evidence and keys files on the given columns
}
\examples{
{
evidenceKeys <- artms_mergeEvidenceAndKeys(data = artms_data_ph_evidence, 
                                           keys = artms_data_ph_keys)
}
}
\keyword{evidence,}
\keyword{keys}
\keyword{merge,}
