% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_mergeEvidenceKeysByFiles}
\alias{artms_mergeEvidenceKeysByFiles}
\title{Merge evidence and keys by file name}
\usage{
artms_mergeEvidenceKeysByFiles(evidence_file, keys_file)
}
\arguments{
\item{evidence_file}{The Evidence file name}

\item{keys_file}{The keys file name}
}
\value{
A data.frame with both evidence and keys files merged by raw.files
}
\description{
Merge evidence and keys by file name
}
\keyword{artms_mergeEvidenceKeysByFiles()}
\keyword{evidence,}
\keyword{keys}
\keyword{merge,}
