% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_mergeMaxQDataWithKeys}
\alias{artms_mergeMaxQDataWithKeys}
\title{Merge evidence and keys files}
\usage{
artms_mergeMaxQDataWithKeys(data, keys, by = c("RawFile"))
}
\arguments{
\item{data}{The evidence in data.frame}

\item{keys}{The keys in data.frame}

\item{by}{Vector specifying the columns use to merge the evidence and keys.}
}
\value{
A new data.frame with the evidence and keys merged
Default: \code{RawFile}
}
\description{
Merge the evidence and keys files on the given columns
}
\keyword{artms_mergeMaxQDataWithKeys()}
\keyword{evidence,}
\keyword{keys}
\keyword{merge,}
