% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotAbundanceBoxplots}
\alias{artms_plotAbundanceBoxplots}
\title{Plot abundance boxplots}
\usage{
artms_plotAbundanceBoxplots(data)
}
\arguments{
\item{data}{modelqc processed data.frame}
}
\value{
Abundacen boxplot
}
\description{
Plot abundance boxplots
}
\keyword{abundance}
\keyword{artms_plotAbundanceBoxplots()}
\keyword{plot,}
