% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotCorrelationConditions}
\alias{artms_plotCorrelationConditions}
\title{Plot correlation between conditions}
\usage{
artms_plotCorrelationConditions(data, numberBiologicalReplicas)
}
\arguments{
\item{data}{Protein Abundance (MSstats modelqc) data.frame}

\item{numberBiologicalReplicas}{Number of biological replicates}
}
\value{
A correlation plot between conditions
}
\description{
Plot correlation between conditions
}
\keyword{artms_plotCorrelationConditions()}
\keyword{correlation}
\keyword{plot,}
