% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotCorrelationDistribution}
\alias{artms_plotCorrelationDistribution}
\title{Plot correlation distributions}
\usage{
artms_plotCorrelationDistribution(MatrixCorrelations)
}
\arguments{
\item{MatrixCorrelations}{Matrix of correlations}
}
\value{
A ggplot2 correlation plot
}
\description{
Plot correlation distributions
}
\keyword{artms_plotCorrelationDistribution()}
\keyword{correlation}
\keyword{plot,}
