% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_plotCorumEnrichment}
\alias{artms_plotCorumEnrichment}
\title{Plot Corum Enrichment results}
\usage{
artms_plotCorumEnrichment(df, outfile, theTitle)
}
\arguments{
\item{df}{(data.frame) output from \code{artms_enrichForComplexes}}

\item{outfile}{(char) output file name (must have the extenstion \code{.pdf})}

\item{theTitle}{(char) Plot's title}
}
\value{
(pdf) heatmap of the significantly enriched protein complexes
}
\description{
Heatmap of significant enrichment
}
\examples{
\donttest{
artms_plotCorumEnrichment(df = "log2fc-results-complexEnrichment.txt", 
             outfile = "log2fc-results-complexEnrichment.pdf",
             theTitle = "Protein Complex Enrichment (log2fc > 1)")
}
}
\keyword{enrichment}
\keyword{heatmap,}
\keyword{plot,}
