% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotHeatmap}
\alias{artms_plotHeatmap}
\title{Outputs a heatmap of the MSStats results created using the log2fold
changes}
\usage{
artms_plotHeatmap(input_file, output_file, labels = "*",
  cluster_cols = F, display = "log2FC", lfc_lower = -2,
  lfc_upper = 2, FDR = 0.05)
}
\arguments{
\item{input_file}{MSstats \code{results.txt} file and location}

\item{output_file}{Output file name (pdf format) and location}

\item{labels}{Vector of uniprot ids if only specific labes would like to
be plotted. Default: all labels}

\item{cluster_cols}{\code{True} or \code{False} to cluster columns. Default: FALSE}

\item{display}{Metric to be displayed (default: \code{log2fc})}

\item{lfc_lower}{Lower limit for the log2fc. Default: -2}

\item{lfc_upper}{Upper limit for the log2fc. Default: +2}

\item{FDR}{Upper limit false discovery rate. Default: 0.05}
}
\value{
A heatmap in pdf format of the MSStats results using the
log2 fold changes and the data used to generate the heatmap
}
\description{
Outputs a heatmap of the MSStats results created using the
log2 fold changes.
}
\keyword{artms_plotHeatmap()}
\keyword{heatmap,}
\keyword{log2fc}
