% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotNumberProteinsAbundance}
\alias{artms_plotNumberProteinsAbundance}
\title{Total Number of unique proteins based on abundance data}
\usage{
artms_plotNumberProteinsAbundance(data)
}
\arguments{
\item{data}{modelqc data.frame}
}
\value{
Barplots with the number of proteins per br / condition
}
\description{
Total Number of unique proteins per biological replicate and
conditions
}
\keyword{abundance,}
\keyword{artms_plotNumberProteinsAbundance()}
\keyword{counts}
\keyword{plots,}
