% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotRatioLog2fc}
\alias{artms_plotRatioLog2fc}
\title{Plot correlation between quantifications different quantified
comparisons}
\usage{
artms_plotRatioLog2fc(datai)
}
\arguments{
\item{datai}{Processed MSstats results}
}
\value{
Plot correlation between quantifications and r values
}
\description{
Plot correlation between all quantifications, i.e., different
quantified comparisons
}
\keyword{artms_plotRatioLog2fc()}
\keyword{correlation,}
\keyword{log2fc}
\keyword{plot,}
