% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotReproducibilityAbundance}
\alias{artms_plotReproducibilityAbundance}
\title{Generate reproducibility plots based on abundance data
(normalized intensities from MSstats modelqc)}
\usage{
artms_plotReproducibilityAbundance(data)
}
\arguments{
\item{data}{Protein abundance data.frame (modelqc)}
}
\value{
Reproducibility plots based on abundance data (normalized intensities)
}
\description{
Generate reproducibility plots based on abundance data
(normalized intensities from MSstats modelqc)
}
\keyword{abundance}
\keyword{artms_plotReproducibilityAbundance()}
\keyword{plot,}
\keyword{reproducibility,}
