% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotReproducibilityEvidence}
\alias{artms_plotReproducibilityEvidence}
\title{Plot reproducibility plots}
\usage{
artms_plotReproducibilityEvidence(data)
}
\arguments{
\item{data}{evidence data.frame}
}
\value{
A reproducibility plot
}
\description{
Plot reproducibility plots
}
\keyword{artms_plotReproducibilityEvidence()}
\keyword{control}
\keyword{plot,}
\keyword{qc,}
\keyword{quality,}
