% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_overlap.R
\name{artms_plotVenn2}
\alias{artms_plotVenn2}
\title{Genererate Venn Diagram for two groups}
\usage{
artms_plotVenn2(m11, m12, m22, pval = NULL)
}
\arguments{
\item{m11}{Length intersection group A and B}

\item{m12}{Length diff group A and B}

\item{m22}{Length diff uniques}

\item{pval}{p-value (default: NULL)}
}
\value{
a plot of a venn diagram
}
\description{
This will create and save a venn digram that displays both
the number of members ber group, but also the percentage of the whole group
}
\keyword{artms_plotVenn2()}
\keyword{diagram}
\keyword{plot,}
\keyword{venn,}
