% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_overlap.R
\name{artms_proportionPlot}
\alias{artms_proportionPlot}
\title{Make a horizontal protein plot of the proportion}
\usage{
artms_proportionPlot(x, out_file)
}
\arguments{
\item{x}{overlap table}

\item{out_file}{out file name}
}
\value{
plots
}
\description{
Make a horizontal protein plot of the proportion
}
\keyword{artms_proportionPlot()}
\keyword{plot,}
\keyword{proportion}
