% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlEvidenceExtended.R
\name{artms_qualityControlEvidenceExtended}
\alias{artms_qualityControlEvidenceExtended}
\title{Extended Quality Control of the MaxQuant evidence.txt file}
\usage{
artms_qualityControlEvidenceExtended(evidence_file, keys_file,
  plotPSM = TRUE, plotIONS = TRUE, plotTYPE = TRUE,
  plotPEPTIDES = TRUE, plotPROTEINS = TRUE, plotPIO = TRUE,
  plotCS = TRUE, plotME = TRUE, plotMOCD = TRUE, plotPEPICV = TRUE,
  plotPEPDETECT = TRUE, plotPROTICV = TRUE, plotPROTDETECT = TRUE,
  plotIDoverlap = TRUE, plotIC = TRUE, plotSP = TRUE)
}
\arguments{
\item{evidence_file}{(char or data.frame) The evidence file path and name, or
data.frame}

\item{keys_file}{(char or data.frame) The keys file path and name or
data.frame}

\item{plotPSM}{(logical) \code{TRUE} generates PSM plots}

\item{plotIONS}{(logical) Plot a QC plot}

\item{plotTYPE}{(logical) Plot a QC plot}

\item{plotPEPTIDES}{(logical) Plot a QC plot}

\item{plotPROTEINS}{(logical) Plot a QC plot}

\item{plotPIO}{(logical) Plot a QC plot}

\item{plotCS}{(logical) Plot a QC plot}

\item{plotME}{(logical) Plot a QC plot}

\item{plotMOCD}{(logical) Plot a QC plot}

\item{plotPEPICV}{(logical) Plot a QC plot}

\item{plotPEPDETECT}{(logical) Plot a QC plot}

\item{plotPROTICV}{(logical) Plot a QC plot}

\item{plotPROTDETECT}{(logical) Plot a QC plot}

\item{plotIDoverlap}{(logical) Plot a QC plot}

\item{plotIC}{(logical) Plot a QC plot}

\item{plotSP}{(logical) Plot a QC plot}
}
\value{
A number of QC plots based on the evidence file
}
\description{
Performs quality control based on the information available in
the MaxQuant \code{evidence.txt} file.
}
\details{
all the plots are generated by default
}
\examples{
\donttest{
artms_qualityControlEvidenceExtended(evidence_file, keys_file)
}
}
\keyword{evidence,}
\keyword{keys}
\keyword{qc,}
