% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{artms_quantification}
\alias{artms_quantification}
\title{Relative quantification using MSstats}
\usage{
artms_quantification(yaml_config_file)
}
\arguments{
\item{yaml_config_file}{(char) The yaml file name and location}
}
\value{
The relative quantification of the conditions and comparisons
specified in the keys/contrast file resulting from running MSstats, in
addition to quality control plots (if selected)
}
\description{
Relative quantification using MSstats including:
\itemize{
\item plots
\item quantifications (log2fc, pvalues, etc)
\item normalized abundance values
}
}
\examples{
\donttest{
artms_quantification("artms-ab-config.yaml")
}
}
\keyword{driver,}
\keyword{function}
\keyword{main,}
