% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_RemoveProtBelowThres}
\alias{artms_RemoveProtBelowThres}
\title{Filter: Remove proteins below some threshold of minimal reproducibility}
\usage{
artms_RemoveProtBelowThres(dfi, mnbr)
}
\arguments{
\item{dfi}{Data.frame with biological replicates information}

\item{mnbr}{minimal number of biological replicates}
}
\value{
a filtered \code{dfi}
}
\description{
If a protein is not found in a minimal number of
biological replicates in at least one of the conditions, it is removed
}
\keyword{artms_RemoveProtBelowThres()}
\keyword{bioreplicates,}
\keyword{filter,}
\keyword{reproducibility}
