% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{artms_resultsWide}
\alias{artms_resultsWide}
\title{Reshape the MSstats results file from long to wide format}
\usage{
artms_resultsWide(input_file, output_file)
}
\arguments{
\item{input_file}{Input file name and location (MSstats \code{results.txt} file)}

\item{output_file}{Output file name and location}
}
\value{
Reshaped file with unique protein ids and as many columns log2fc
and adj.pvalues as comparisons available
for as many
}
\description{
Converts the normal MSStats output file into "wide" format
where each row represents a protein's results, and each column represents
the comparison made by MSStats. The fold change and p-value of each
comparison will be it's own column.
}
\keyword{artms_resultsWide()}
