% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSstats.R
\name{artms_runMSstats}
\alias{artms_runMSstats}
\title{Run MSstats}
\usage{
artms_runMSstats(dmss, contrasts, config)
}
\arguments{
\item{dmss}{Formatted and filtered evidence file (MSstats format)}

\item{contrasts}{The contrast data.frame in MSstats format}

\item{config}{the configation (imported yaml) object}
}
\value{
It generates several output files
\itemize{
\item If selected \code{before} and/or \code{after}, the \code{ProfilePlot} and \code{QCPlot} plots
by the MSstats \code{dataProcessPlots} function are generated
(in \code{.pdf} format)
\item Text file output of \code{quantification()} (\code{-mss-sampleQuant.txt})
\item Text file output of \code{quantification(type="Group")} (\code{-mss-groupQuant.txt})
\item MSstats \code{ProcessedData} normalized results (\code{-mss-normalized.txt})
\item MSstats \code{ComparisonResult} results
\item MSstats \code{ModelQC} results
\item MSstats \code{designSampleSize} sample size
\item MSstats \code{designSampleSize} power experiment
}
}
\description{
Run MSstats giving a processed evidence file (MSstats format)
and a contrast file. It also generates a series of summary plots before,
and after normalization.
}
\keyword{MSstats,}
\keyword{QC}
\keyword{artms_runMSstats()}
\keyword{contrast,}
\keyword{intensity,}
\keyword{plots,}
\keyword{run,}
