% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_samplePeptideBarplot}
\alias{artms_samplePeptideBarplot}
\title{Barplot of peptide counts per biological replicate}
\usage{
artms_samplePeptideBarplot(data_f, config)
}
\arguments{
\item{data_f}{Evidence file (same structure as the original)}

\item{config}{Configuration object}
}
\value{
Barplot of peptide counts
}
\description{
Total number of unique peptide identified per biological
replicate
}
\keyword{artms_samplePeptideBarplot()}
\keyword{barplot,}
\keyword{counts,}
\keyword{peptides}
