% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_selectTheOneLog2fc}
\alias{artms_selectTheOneLog2fc}
\title{Select and label the condition more abundant in a quantification}
\usage{
artms_selectTheOneLog2fc(a, b)
}
\arguments{
\item{a}{log2fc column}

\item{b}{comparison column}
}
\value{
One of the conditions from the comparison
}
\description{
Select and label the condition more abundant in a quantification
\itemize{
\item If log2fc > 0 the condition on the left ('numerator') is the most abundant
\item If log2fc < 0 the condition on the right ('denominator') is the most abundant
}
}
\keyword{artms_selectTheOneLog2fc()}
\keyword{internal,}
\keyword{labeling}
\keyword{selection,}
