% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_SILACtoLong}
\alias{artms_SILACtoLong}
\title{Convert the SILAC evidence file to MSstats format}
\usage{
artms_SILACtoLong(filename, output)
}
\arguments{
\item{filename}{Text filepath to the evidence file}

\item{output}{Text filepath of the output name}
}
\value{
A data.frame with SILAC data processed for MSstats
}
\description{
Converting the evidence file from a SILAC search to a format
compatible with MSstats. It basically modifies the Raw.files adding the
Heavy and Light label
}
\keyword{artms_SILACtoLong()}
