% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_spectralCounts}
\alias{artms_spectralCounts}
\title{Outputs the spectral counts from the MaxQuant evidence file.}
\usage{
artms_spectralCounts(input_file, keys_file, output_file)
}
\arguments{
\item{input_file}{Maxquant evidence file}

\item{keys_file}{Keys file with the experimental design}

\item{output_file}{Output file name (add \code{.txt} extenstion)}
}
\value{
A txt file with biological replicates, protein id, and spectral
count columns
}
\description{
Outputs the spectral counts from the MaxQuant evidence file.
}
\keyword{artms_spectralCounts()}
\keyword{evidence}
\keyword{spectral_counts,}
