% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{artms_writeConfigYamlFile}
\alias{artms_writeConfigYamlFile}
\title{Write out a template file of the artMS configuration file (yaml)}
\usage{
artms_writeConfigYamlFile(config_file_name = "artms_config_file.yaml")
}
\arguments{
\item{config_file_name}{(char) The name for the configuration file. It must
have a \code{.yaml} extension. If \code{NULL}, it returns the config as a yaml object}
}
\value{
A file (or yaml data object) of the artMS configuration file
}
\description{
Creates a template file of the artMS configuration file, which
is required to run \code{artms_quantification}. Check \code{?artms_config} and the
vignettes to find out more about the details of the structure of the file
and how to fill it up
}
\examples{
config_empty <- artms_writeConfigYamlFile(config_file_name = NULL)
}
\keyword{config,}
\keyword{yaml}
