% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artmsFilterEvidenceContaminants}
\alias{artmsFilterEvidenceContaminants}
\title{Remove contaminants and empty proteins from the MaxQuant evidence file}
\usage{
artmsFilterEvidenceContaminants(x, verbose = TRUE)
}
\arguments{
\item{x}{(data.frame) of the Evidence file}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
(data.frame) without REV__ and CON__ Protein ids
}
\description{
Remove contaminants and erronously identified 'reverse'
sequences by MaxQuant, in addition to empty protein ids
}
\examples{
ef <- artmsFilterEvidenceContaminants(x = artms_data_ph_evidence)
}
\keyword{cleanup,}
\keyword{contaminants}
