% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinToSiteConversion.R
\name{artmsProtein2SiteConversion}
\alias{artmsProtein2SiteConversion}
\title{Converts the \code{Proteins} column of the evidence file to site-specific
\code{Uniprot_PTM} notation}
\usage{
artmsProtein2SiteConversion(evidence_file, ref_proteome_file, output_file,
  mod_type, verbose = TRUE)
}
\arguments{
\item{evidence_file}{(char) The evidence file name and location}

\item{ref_proteome_file}{(char) The reference proteome used as database
to search the \code{evidence.txt} file with MaxQuant. It will be used to map the
modified peptide to the protein sequence and find the site location.
Therefore, it does not use the MaxQuant's \code{Phospho (STY)Sites.txt}}

\item{output_file}{(char) Output file name
(\code{-sites-evidence.txt} recommended)}

\item{mod_type}{(char) The posttranslational modification. Options:
\itemize{
\item \code{UB}: Protein Ubiquitination
\item \code{PH}: Protein Phosphorylation
\item \code{AC}: Protein Acetylation
}}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
(file) Return a new evidence file with the \code{Proteins} column
modified by adding the sequence site location(s) + postranslational
modification(s) to the uniprot entry / refseq id.

Examples: \code{A34890_ph3}; \code{Q64890_ph24_ph456}; \code{Q64890_ub34_ub129_ub234};
\code{Q64890_ac35}.
}
\description{
It enables the site-specific quantification of PTMs by
converting the \code{Proteins} column of the evidence file to an \code{Uniprot_PTM}
or \code{RefSeq_PTM} notation.
In this way, each of the modified peptides can be quantified
independently across conditions.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
WARNING: we have detected a version of MaxQuant (1.6.3.4) outputs a`
"Modified sequence" column of the evidence file that has two important
changes for the annotation of phosphorylation:
\itemize{
\item Uses \code{p} instead of \code{(ph)}
\item The modified residue (i.e. STY) is the residue on the left of the \code{p},
instead of the residue to the right of \code{(ph)} as usual.
These changes significantly affect the algorithm used by this artMS function
to extract the PTM site information. We are working on debugging the issue
and use an alternative function independent of the "Mofified sequence"
column. If your MaxQuant version is affected, then the column \code{mod_sites} of
the output \code{-mapping.txt} will have many NA values for peptides with \code{p}
modifications (and no NA values should be present)
}
}
\examples{
# Testing warning if files are not submitted
artmsProtein2SiteConversion(evidence_file = NULL, ref_proteome_file = NULL, 
output_file = NULL)
}
\keyword{ac}
\keyword{convert,}
\keyword{evidence,}
\keyword{ph,}
\keyword{ptm,}
\keyword{ub,}
