% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlSummaryExtended.R
\name{artmsQualityControlSummaryExtended}
\alias{artmsQualityControlSummaryExtended}
\title{Quality Control of the MaxQuant summary.txt file}
\usage{
artmsQualityControlSummaryExtended(summary_file, keys_file,
  isFractions = FALSE, plotMS1SCANS = TRUE, plotMS2 = TRUE,
  plotMSMS = TRUE, plotISOTOPE = TRUE, verbose = TRUE)
}
\arguments{
\item{summary_file}{(char or data.frame) The evidence file path and name, or
data.frame}

\item{keys_file}{(char or data.frame) The keys file path and name or
data.frame}

\item{isFractions}{(logical) \code{TRUE} if it is a 2D experiment (fractions).
Default: \code{FALSE}}

\item{plotMS1SCANS}{(logical) It plots whatever}

\item{plotMS2}{(logical) It plots whatever}

\item{plotMSMS}{(logical) It plots whatever}

\item{plotISOTOPE}{(logical) It plots whatever}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
A number of plots from the summary file
}
\description{
Performs quality control based on the information available in
the MaxQuant summary.txt file.
}
\examples{
# Testing warning if files are not submitted
test <- artmsQualityControlSummaryExtended(summary_file = NULL,
keys_file = NULL)
}
\keyword{keys}
\keyword{qc,}
\keyword{summary,}
