% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{castMaxQToWide}
\alias{castMaxQToWide}
\title{Long to Wide format using the \code{Sequence} column of the evidence file}
\usage{
castMaxQToWide(d_long)
}
\arguments{
\item{d_long}{the data.frame in long format}
}
\value{
Evidence file reshaped by rawfile and IsotopeLabelType
}
\description{
Facilitates applying the dcast function, i.e., takes long-format
data and casts it into wide-format data.
}
\keyword{castMaxQToWide()}
\keyword{data.frame,}
\keyword{dcast}
