% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{castMaxQToWidePTM}
\alias{castMaxQToWidePTM}
\title{Long to Wide format selecting the \code{Modified.sequence} column of the
evidence file}
\usage{
castMaxQToWidePTM(d_long)
}
\arguments{
\item{d_long}{the data.frame in long format}
}
\value{
Evidence file reshaped by rawfile and IsotopeLabelType
}
\description{
Facilitates applying the dcast function, i.e., takes long-format
data and casts it into wide-format data.
}
\keyword{castMaxQToWidePTM()}
\keyword{data.frame,}
\keyword{dcast}
