% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{changeColumnName}
\alias{changeColumnName}
\title{Change a specific column name in a given data.frame}
\usage{
changeColumnName(dataset, oldname, newname)
}
\arguments{
\item{dataset}{the data.frame with the column name you want to change}

\item{oldname}{the old column name}

\item{neename}{the new name for that column}
}
\value{
A new data.frame with the new specified column name
}
\description{
Making easier j
}
\keyword{changeColumnName()}
\keyword{columns}
\keyword{data.frame,}
\keyword{rename,}
