% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfFile.R
\name{.artms_checkIfFile}
\alias{.artms_checkIfFile}
\title{Check if an input is a file or a data object}
\usage{
.artms_checkIfFile(input_file, is.evidence = FALSE)
}
\arguments{
\item{input_file}{(object or data.frame) The filepath/object to be checked.}

\item{is.evidence}{(logical) Whether or not the file to be read in is an
evidence file. This will assign proper classes to the evidence file
when being read in.}
}
\value{
An R data object
}
\description{
This function is used in order to make it so a user can submit
either a path to a data file or a data object in data.frame or data.table
form.
}
\keyword{evidence,}
\keyword{file,}
\keyword{input}
\keyword{internal,}
