% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_checkRawFileColumnName}
\alias{.artms_checkRawFileColumnName}
\title{Check the \code{Raw file} column name on the evidence or keys data.frame}
\usage{
.artms_checkRawFileColumnName(df)
}
\arguments{
\item{(data.frame)}{keys or evidence files}
}
\value{
(data.frame) with the \code{RawFile} column name
}
\description{
Depending on how the data is loaded, the \code{Raw file} column
might have different format. This function check to ensure consistency in
both the evidence and keys data.frames
}
\keyword{.artms_checkRawFileColumnName()}
\keyword{columname}
\keyword{internal}
\keyword{rawfile,}
