% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{.artms_cleanGPROFILER}
\alias{.artms_cleanGPROFILER}
\title{Simplify the gProfiler output}
\usage{
.artms_cleanGPROFILER(gp)
}
\arguments{
\item{gp}{(data.frame) with the results}
}
\value{
(data.frame) with the following columns:
'query.number', 'domain', 'p.value', 'query.size', 'overlap.size',
'term.size', 'recall', 'precision', 'term.id', 'term.name',
'intersection'
}
\description{
Simplify the output from \code{artms_enrichProfiler} resulted from
running \code{gProfileR}
}
\keyword{cleaning}
\keyword{internal,}
