% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{.artms_EnrichmentPlotHeatmaps}
\alias{.artms_EnrichmentPlotHeatmaps}
\title{Plot and save heatmaps of the significant enrichment results}
\usage{
.artms_EnrichmentPlotHeatmaps(dat, out_file)
}
\arguments{
\item{dat}{(data.frame) output from gprofiler}

\item{out_file}{(char) output file name (must have \code{.txt} extension)}
}
\value{
(pdf) A heatmap of the most significant enrichments
}
\description{
plot and save heatmaps of the significant enrichment results
}
\keyword{.artms_EnrichmentPlotHeatmaps()}
\keyword{enrichments}
\keyword{heatmap,}
\keyword{internal,}
\keyword{plot,}
