% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_filterData}
\alias{.artms_filterData}
\title{Filtering data}
\usage{
.artms_filterData(data, config)
}
\arguments{
\item{data}{(data.frame) Evidence file}

\item{config}{(yaml.object) Configuration object (opened yaml file)}
}
\description{
Apply the filtering options, i.e., remove protein groups and/or
contaminants, and/or, select posttranslational modification (if any)
}
\keyword{.artms_filterData()}
\keyword{filtering,}
\keyword{internal,}
\keyword{proteingroups,}
\keyword{ptms}
\keyword{remove,}
