% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{.artms_foldComplexEnrichment}
\alias{.artms_foldComplexEnrichment}
\title{Enrichment analysis of Protein Complexes (based on CORUM database)}
\usage{
.artms_foldComplexEnrichment(mylist, corum, background)
}
\arguments{
\item{mylist}{(vector) of protein ids}

\item{corum}{(data.frame) The corum database (with the corum format
and labels)}

\item{background}{(int) Total number of proteins (number) to use as
background}
}
\value{
(data.frame) with the list of protein complexes with a fold change
larger than 1
}
\description{
Enrichment analysis of Protein Complexes
(based on CORUM database)
}
\keyword{.artms_foldComplexEnrichment()}
\keyword{complexes}
\keyword{enrichment,}
\keyword{internal,}
\keyword{protein,}
