% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSstatsformat.R
\name{.artms_getMSstatsFormat}
\alias{.artms_getMSstatsFormat}
\title{Generate MSstats format object and file}
\usage{
.artms_getMSstatsFormat(data_f, fraction, datafile, funfunc = "sum")
}
\arguments{
\item{data_f}{(data.frame) of the filtered Maxquant evidence file.}

\item{fraction}{(boolean) 1 or 0 option to specified whether or not
is a fractionated experiment}

\item{datafile}{(char) The evidence file name (to generate the output file)}

\item{funfunc}{(char) The function to use to aggregating the data if it is a
fractionated experiment (default: \code{sum})}
}
\value{
(data.frame) MSstats compatible format
}
\description{
Takes as input a reduced version of the Evidence file and
generates the input data.frame required by MSstats.
It processes fractionated data as well.
}
\keyword{MSstats,}
\keyword{format,}
\keyword{fractions}
\keyword{input,}
\keyword{internal,}
