% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_getPCAplots}
\alias{.artms_getPCAplots}
\title{Generate PCA plots based on abundance data}
\usage{
.artms_getPCAplots(data, filename, allConditions)
}
\arguments{
\item{data}{Data.frame output from \code{artms_loadModelQCstrict}}

\item{filename}{Prefix to generate output names (WITH NO EXTENSION)}

\item{allConditions}{Conditions selected to generate the plots}
}
\value{
PCA plots based on abundance data (pdf format)
}
\description{
Generate PCA plots based on abundance data
}
\keyword{internal,}
\keyword{pca}
\keyword{plot,}
