% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlEvidenceExtended.R
\name{.artms_panelCor}
\alias{.artms_panelCor}
\title{Panel Cor}
\usage{
.artms_panelCor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{(datatype) Describe}

\item{y}{(datatype) Describe}

\item{digits}{A number representing...}

\item{prefix}{A prefix for the...}

\item{cex.cor}{A Cex cor....}

\item{...}{No idea what this does}
}
\value{
What does it return?
}
\description{
Describe
}
\keyword{internal,}
\keyword{stats}
