% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlEvidenceExtended.R
\name{.artms_panelHist}
\alias{.artms_panelHist}
\title{Panel Histogram}
\usage{
.artms_panelHist(x, ...)
}
\arguments{
\item{x}{(datatype) Describe}

\item{...}{No idea what this does}
}
\value{
What does it return?
}
\description{
Describe
}
\keyword{internal,}
\keyword{stats}
