% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_plotAbundanceBoxplots}
\alias{.artms_plotAbundanceBoxplots}
\title{Plot abundance boxplots}
\usage{
.artms_plotAbundanceBoxplots(data)
}
\arguments{
\item{data}{(data.frame) processed modelqc}
}
\value{
Abundacen boxplot
}
\description{
Plot abundance boxplots
}
\keyword{abundance}
\keyword{internal,}
\keyword{plot,}
