% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_plotCorrelationConditions}
\alias{.artms_plotCorrelationConditions}
\title{Plot correlation between conditions}
\usage{
.artms_plotCorrelationConditions(data, numberBiologicalReplicas)
}
\arguments{
\item{data}{(data.frame) of Protein Abundance (MSstats modelqc)}

\item{numberBiologicalReplicas}{(int) Number of biological replicates}
}
\value{
(ggplot.object) A correlation plot between conditions
}
\description{
Plot correlation between conditions
}
\keyword{correlation}
\keyword{internal,}
\keyword{plot,}
