% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{.artms_plotCorumEnrichment}
\alias{.artms_plotCorumEnrichment}
\title{Plot Corum Enrichment results}
\usage{
.artms_plotCorumEnrichment(df, outfile, theTitle)
}
\arguments{
\item{df}{(data.frame) output from \code{artms_enrichForComplexes}}

\item{outfile}{(char) output file name (must have the extenstion \code{.pdf})}

\item{theTitle}{(char) Plot's title}
}
\value{
(pdf) heatmap of the significantly enriched protein complexes
}
\description{
Heatmap of significant enrichment
}
\keyword{enrichment}
\keyword{heatmap,}
\keyword{internal,}
\keyword{plot,}
