% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_plotHeat}
\alias{.artms_plotHeat}
\title{Heatmap of significant values}
\usage{
.artms_plotHeat(mss_F, out_file, labelOrder = NULL, names = "Protein",
  cluster_cols = FALSE, display = "log2FC")
}
\arguments{
\item{mss_F}{(data.frame) with the significant values (log2fc, pvalues)}

\item{out_file}{(char) Name for the output}

\item{labelOrder}{(vector) Vector with the particular order for the IDs
(default, \code{NULL} no order)}

\item{names}{(char) Type of ID used. Default is \code{Protein} (uniprot entry id).
Soon will be possible to use 'Gene' name ids.}

\item{cluster_cols}{(logical) Select whether to cluster the columns.
Options: \code{T} or \code{F}. Default \code{T}.}

\item{display}{(char) Value used to genarate the heatmaps. Options:
\itemize{
\item \code{log2FC} (default)
\item \code{adj.pvalue}
\item \code{pvalue}
}}
}
\value{
A heatmap of significant values
}
\description{
heatmap plot to represent proteins with significant changes
}
\keyword{heatmap}
\keyword{significant,}
