% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_plotReproducibilityAbundance}
\alias{.artms_plotReproducibilityAbundance}
\title{Generate reproducibility plots based on abundance data
(normalized intensities from MSstats modelqc)}
\usage{
.artms_plotReproducibilityAbundance(data)
}
\arguments{
\item{data}{(data.frame) Protein abundance (modelqc)}
}
\value{
Reproducibility plots based on abundance data (normalized intensities)
}
\description{
Generate reproducibility plots based on abundance data
(normalized intensities from MSstats modelqc)
}
\keyword{abundance}
\keyword{plot,}
\keyword{reproducibility,}
