% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.artms_plotReproducibilityEvidence}
\alias{.artms_plotReproducibilityEvidence}
\title{Generate reproducibility plots based on raw intentities
(from the evidence file)}
\usage{
.artms_plotReproducibilityEvidence(data)
}
\arguments{
\item{data}{(data.frame) clean processed evidence}
}
\value{
(pdf) A reproducibility plot based on evidence values
}
\description{
Generate reproducibility plots based on raw intentities
(log tranformed) from the evidence file
}
\keyword{control}
\keyword{internal,}
\keyword{plot,}
\keyword{qc,}
\keyword{quality,}
