% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_prettyPrintHeatmapLabels}
\alias{.artms_prettyPrintHeatmapLabels}
\title{Pretty Labels for Heatmaps}
\usage{
.artms_prettyPrintHeatmapLabels(uniprot_acs, uniprot_ids, gene_names)
}
\arguments{
\item{uniprot_acs}{(char) Uniprot accession id}

\item{uniprot_ids}{(char) Uniprot entry id}

\item{gene_names}{(car) Gene symbol}
}
\value{
Pretty labels for a heatmap
}
\description{
Generates pretty labels for the heatmaps.
}
\keyword{.artms_prettyPrintHeatmapLabels()}
\keyword{internal,}
\keyword{plots,}
\keyword{pretty}
