% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlEvidenceExtended.R
\name{.artms_qcut}
\alias{.artms_qcut}
\title{Quantiles}
\usage{
.artms_qcut(x, n)
}
\arguments{
\item{x}{(datatype) Describe}

\item{n}{(datatype) Describe}
}
\value{
What does it return?
}
\description{
Quantiles
}
\keyword{internal,}
\keyword{stats}
