% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfFile.R
\name{.artms_read_evidence_file}
\alias{.artms_read_evidence_file}
\title{Read the Evidence File}
\usage{
.artms_read_evidence_file(evidence_file)
}
\arguments{
\item{evidence_file}{(char)The filepath to the MaxQuant searched data
(evidence) file (txt tab delimited file).}
}
\value{
(data.frame) with the evidence file with defining classes
}
\description{
Read in a MaxQuant searched Evidence file using data.table.
This function properly classes each column and so fread doesn't have
to guess.
}
\keyword{.artms_read_evidence_file()}
\keyword{MaxQuant,}
\keyword{evidence}
