% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_removeMaxQProteinGroups}
\alias{.artms_removeMaxQProteinGroups}
\title{Remove protein groups}
\usage{
.artms_removeMaxQProteinGroups(data)
}
\arguments{
\item{data}{Data.frame with a \code{Proteins} column.}
}
\value{
A data.frame with the protein groups removed
}
\description{
Remove the group of proteins ids separated by separated by \code{;}
}
\keyword{.artms_removeMaxQProteinGroups()}
\keyword{maxquant,}
\keyword{proteingroups}
\keyword{remove,}
