% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_sampleCorrelationHeatmap}
\alias{.artms_sampleCorrelationHeatmap}
\title{Correlation heatmaps of all the individual features}
\usage{
.artms_sampleCorrelationHeatmap(data_w, keys, config)
}
\arguments{
\item{data_w}{(data.frame) resulting from the \code{.artms_castMaxQToWidePTM}
function}

\item{keys}{(data.frame) of the keys}

\item{config}{(yaml.object) Configuration object (yaml loaded)}
}
\value{
(pdf) A correlation heatmap (suffix \code{-heatmap.pdf})
}
\description{
Correlation heatmap using intensity values across all the
conditions
}
\keyword{comparisons}
\keyword{heatmap,}
\keyword{intensity,}
\keyword{internal,}
