% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlEvidenceExtended.R
\name{.artms_sampleOverlap}
\alias{.artms_sampleOverlap}
\title{Sample overlap}
\usage{
.artms_sampleOverlap(data, sampleID = "bioreplicate",
  referenceID = "Sequence")
}
\arguments{
\item{data}{(datatype) Describe}

\item{sampleID}{(datatype) Describe}

\item{referenceID}{(datatype) Describe}
}
\value{
What does it return?
}
\description{
Describe
}
\keyword{internal,}
\keyword{stats}
