% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_significantHits}
\alias{.artms_significantHits}
\title{Select significant hits}
\usage{
.artms_significantHits(mss_results, labels = "*", LFC = c(-2, 2),
  whatPvalue = "adj.pvalue", FDR = 0.05)
}
\arguments{
\item{mss_results}{(data.frame) of MSstats results}

\item{labels}{(vector) of selected labels. Default: all (\code{*})}

\item{LFC}{(vector, int) with the negative and positive threshold. Default:
c(-2, 2)}

\item{whatPvalue}{(char) \code{pvalue} or \code{adj.pvalue} (default)?}

\item{FDR}{(int) false discovery rate (adj.pvalue) threshold. Default: 0.05}
}
\value{
(data.frame) only with significant hits
}
\description{
Filtered data.frame with significant values (log2fc > 2 |
log2fc < -2; adj.pvalue < 0.05) from the MSstats results
}
\keyword{.artms_significantHits()}
\keyword{internal,}
\keyword{selections}
\keyword{significant,}
