% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_trim}
\alias{.artms_trim}
\title{Remove white spaces}
\usage{
.artms_trim(x)
}
\arguments{
\item{x}{(vector) A string}
}
\description{
Remove white spaces
}
\keyword{internal}
\keyword{remove,}
\keyword{whitespace}
