% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfFile.R
\name{.artms_unifyColumnNames}
\alias{.artms_unifyColumnNames}
\title{Unify column names}
\usage{
.artms_unifyColumnNames(x)
}
\arguments{
\item{x}{(data.frame) Loaded evidence file.}
}
\value{
(data.frame) with the evidence file with defining classes and
unified columns
}
\description{
The newer versions of MaxQuant change the letter case in the
column names. This is to help keep things unified still.
}
\keyword{.artms_unifyColumnNames()}
\keyword{MaxQuant,}
\keyword{evidence}
\keyword{internal,}
