% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{.artms_writeContrast}
\alias{.artms_writeContrast}
\title{Generate the contrast matrix required by MSstats from a txt file}
\usage{
.artms_writeContrast(contrast_file, all_conditions = NULL)
}
\arguments{
\item{contrast_file}{The text filepath of contrasts}

\item{all_conditions}{a vector with all the conditions in the keys file}
}
\value{
(data.frame) with the contrast file in the format required by
MSstats
}
\description{
It simplifies the process of creating the contrast file
}
\author{
Tom Nguyen, David Jimenez-Morales
}
\keyword{check,}
\keyword{contrast}
