% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExtras.R
\name{.artms_writeExtras}
\alias{.artms_writeExtras}
\title{Write extras}
\usage{
.artms_writeExtras(results, config)
}
\arguments{
\item{results}{MSstats results}

\item{config}{The configuration object (yaml)}
}
\value{
Extras as selected in the yaml file, including:
\itemize{
\item volcano plot (pdf)
\item Adding annotations (gene symbol based on uniprot)
}
}
\description{
Extras after MSstats, as annotations, volcano plots, heatmaps
}
\keyword{annotations,}
\keyword{extras,}
\keyword{volcano}
