% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExtras.R
\name{.extras_annotate}
\alias{.extras_annotate}
\title{Annotate the files based on the Uniprot accession id}
\usage{
.extras_annotate(results, output_file, uniprot_ac_col = "Protein",
  group_sep = ";", uniprot_dir = "~/github/kroganlab/source/db/",
  species = "HUMAN")
}
\arguments{
\item{results}{MSstats results}

\item{output_file}{output file name}

\item{uniprot_ac_col}{Column with uniprot ids}

\item{group_sep}{Group separation character (if any. Eg: \code{;})}

\item{uniprot_dir}{Directory with the Uniprot mappings}

\item{species}{Species (dash separated accordind to the uniprot file name)}
}
\value{
Annotated data.frame
}
\description{
Annotate the files based on the Uniprot accession id
}
\keyword{annotations}
\keyword{extras,}
\keyword{internal,}
