% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ref.R
\name{.select_ref}
\alias{.select_ref}
\title{Select Reference protein(s) for normalization

NEEDS TO BE FIXED}
\usage{
.select_ref(keys_file, dat_file)
}
\arguments{
\item{keys_file}{(char) The filepath to the MSStats formatted keys file
(txt tab delimited file).}

\item{dat_file}{(char) The filepath to the MaxQuant searched data (evidence)
file (txt tab delimited file).}
}
\value{
(shiny) app to choose a protein to be use for normalization
}
\description{
This function creates an interactive environment where the
user can select which prein(s) to use as a reference when normalizing
the different samples prior to analysis. Only proteins that are present
in each sample are considered. The user can remove the unnecessary/unwanted
proteins by checking the box next to the protein name. One can use the
general view of the proteins plotted across the samples, or at a clustering
of the proteins.
}
\keyword{internal,}
\keyword{msstat,}
\keyword{normalization}
\keyword{protein,}
\keyword{reference}
