% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{filterData}
\alias{filterData}
\title{Filtering data}
\usage{
filterData(data, config)
}
\arguments{
\item{data}{Evidence file (data.frame)}

\item{config}{Configuration object (opened yaml file)}
}
\description{
Apply the filtering options, i.e., remove protein groups and/or
contaminants, and/or, select posttranslational modification (if any)
}
\keyword{filterData()}
\keyword{filtering,}
\keyword{proteingroups,}
\keyword{ptms}
\keyword{remove,}
