% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{filterMaxqData}
\alias{filterMaxqData}
\title{Remove contaminants and empty proteins}
\usage{
filterMaxqData(data)
}
\arguments{
\item{data}{the data.frame in long format}
}
\value{
A new data.frame without REV__ and CON__ Proteins
}
\description{
Remove contaminants and erronously identified 'reverse'
sequences by MaxQuant
}
\keyword{cleanup,}
\keyword{contaminants}
\keyword{filterMaxqData()}
