% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOnNumUnique.R
\name{filterNumUniqPep}
\alias{filterNumUniqPep}
\title{Filter MaxQuant data based on Number of Unique Peptides}
\usage{
filterNumUniqPep(evidence_file, keys_file, prot_groups_file, min_pep,
  verbose = TRUE, return.results = FALSE)
}
\arguments{
\item{evidence_file}{The filepath to the MaxQuant searched data (evidence) file (txt tab delimited file).}

\item{keys_file}{The filepath to the keys file used with MSStats (txt tab delimited file). The condition names must match the Condition names provided to the proteinGroups file replicates.}

\item{prot_groups_file}{The filepath to the MaxQuant searched Protein Groups file (txt tab delimited file).}

\item{min_pep}{The minimum number of peptides that need to be found with at least one bait-prey protein pair to consider it a real hit. Anything not meeting this criteria is removed before scoring.}

\item{verbose}{Whether or not to display status reports. Default = TRUE}

\item{return.results}{Whether or not to return the results to the R environment upon completion. This is useful if this is being used in an R pipeline and you want to feed the results directly into the next stage of analysis via an R environment/terminal. Regardless, the results will be written to file. Default = FALSE}
}
\description{
This will remove all condition-protein instance where the number of unique peptides is below the `min_pep` threshold. The results are written out to a file that is named similar to the evidence file, but also contains the filtering cutoff used.
}
\keyword{MSStats}
\keyword{MaxQuant,}
\keyword{evidence,}
\keyword{filterNumUniqPep()}
