% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSstatsformat.R
\name{getMSstatsFormat}
\alias{getMSstatsFormat}
\title{Generate MSstats format object and file}
\usage{
getMSstatsFormat(data_f, fraction, datafile, funfunc)
}
\arguments{
\item{data_f}{Maxquant evidence data.frame filtered}

\item{fraction}{1 or 0 option to specified whether it is a fractionated
experiment}

\item{datafile}{The evidence file name (to generate the output file)}

\item{funfunc}{The function to use to aggregating the data if it is a
fractionated experiment (\code{sum} recommended)}
}
\description{
Takes as input a reduced version of the Evidence file and
generates the input data.frame required by MSstats. It processes fractionated
data.
}
\keyword{MSstats,}
\keyword{format,}
\keyword{fractions}
\keyword{getMSstatsFormat()}
\keyword{input,}
