% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_main.R
\name{main}
\alias{main}
\title{Main Function}
\usage{
main(opt)
}
\arguments{
\item{opt}{the object with all the options}
}
\value{
All the selected options
}
\description{
Main function running all the selected options
}
\keyword{driver,}
\keyword{function}
\keyword{main()}
\keyword{main,}
