% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_main.R
\name{mainART}
\alias{mainART}
\title{Main Function}
\usage{
mainART(yaml_config_file)
}
\arguments{
\item{yaml_config_file}{The yaml file name and location}
}
\value{
All the selected options
}
\description{
Main function running all the selected options
}
\keyword{driver,}
\keyword{function}
\keyword{main,}
\keyword{mainART()}
