% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{mapUniprot2entrezGeneName}
\alias{mapUniprot2entrezGeneName}
\title{Map UniprotKB to Entrez and Gene Name}
\usage{
mapUniprot2entrezGeneName(theUniprots, specie)
}
\arguments{
\item{theUniprots}{List of UniprotKB IDs}

\item{specie}{The specie name. Currently supporting:
\itemize{
\item ANOPHELES
\item ARABIDOPSIS
\item BOVINE
\item WORM
\item CANINE
\item FLY
\item ZEBRAFISH
\item ECOLI_STRAIN_K12
\item ECOLI_STRAIN_SAKAI
\item CHICKEN
\item HUMAN
\item MOUSE
\item RHESUS
\item MALARIA
\item CHIMP
\item RAT
\item PIG
\item XENOPUS
}}
}
\value{
A data.frame with EntrezID and GENENAMES mapped on UniprotKB ids
}
\description{
Map UniprotKB to Entrez and Gene Name
}
\keyword{annotation,}
\keyword{ids}
\keyword{mapUniprot2entrezGeneName()}
