% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_main.R
\name{MQ.msstats}
\alias{MQ.msstats}
\title{Run MSStats pipeline designed for MaxQuant evidence files.}
\usage{
MQ.msstats(opt)
}
\arguments{
\item{opt}{Pick which quantitative value to use: "spectral_counts" or "ms1"}
}
\description{
This function will take a MaxQuant file, keys, contrast, and config file, and handle all the file processing and conversions before running MSStats on the data. This is designed using the MSStats constraints/options outlined in the config.yaml file.
}
\keyword{MQ.msstats()}
\keyword{MSStats}
\keyword{MaxQuant}
