% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq2mss_PTM.R
\name{mq2mss_PTM}
\alias{mq2mss_PTM}
\title{MaxQuant PTM to MSStats.}
\usage{
mq2mss_PTM(ptmsites_file)
}
\arguments{
\item{ptmsites_file}{The filepath/object to the PTM Sites.}
}
\description{
Converts a MaxQuant PTM Sites file into a MSStats compatible
file.
}
\keyword{PTM,}
\keyword{STY}
\keyword{evidence,}
\keyword{file,}
\keyword{input,}
\keyword{modification,}
\keyword{mq2mss_PTM()}
\keyword{post}
\keyword{sites,}
\keyword{translational}
