% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.annotate}
\alias{MQutil.annotate}
\title{Annotate MSStats results file.}
\usage{
MQutil.annotate(input_file = opt$input, output_file = opt$output,
  uniprot_ac_col = "Protein", group_sep = ";",
  uniprot_dir = "~/Box Sync/db/mist/", species = "HUMAN")
}
\arguments{
\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}

\item{output_file}{The filepath to the intended output file (txt tab delimited file).}

\item{uniprot_ac_col}{Name of the column containing the uniprot ID's. Defaults to "Protein".}

\item{uniprot_dir}{Directory location of where the uniprot database flatfiles are located.}

\item{species}{The species contained in this search of uniprot files. Can include multiple species separated by a "-" .}
}
\description{
Annotates the proteins in the results or results-wide files after they've been through the MSStats pipeline. Multiple species can be searched at once, simply separate them by a "-". (eg. human-mouse).
}
\keyword{MQutil.annotate()}
\keyword{annotate}
\keyword{annotation}
\keyword{uniprot}
