% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.concat}
\alias{MQutil.concat}
\title{Concatenate MaxQuant search result files}
\usage{
MQutil.concat(filenames, output)
}
\arguments{
\item{filenames}{The filepaths to the MaxQuant searched results files (txt tab delimited file).}

\item{output}{The filepath to the intended output file (txt tab delimited file).}
}
\description{
Concatenates multiple MaxQuant files into one  big file.
}
\keyword{MQutil.concat()}
\keyword{concatenate}
\keyword{join}
\keyword{merge}
