% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.mapSitesBack}
\alias{MQutil.mapSitesBack}
\title{Map the results back into sites format after MSStats PTM site analysis.}
\usage{
MQutil.mapSitesBack(input_file, mapping_file, output_file)
}
\arguments{
\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}

\item{mapping_file}{The filepath to the mapping file (ending in "-mapping.txt")}

\item{output_file}{The filepath to the intended output file (txt tab delimited file).}
}
\description{
Used with PTM datasets. Map back the sites to correct proteins after MSStats analysis. This file is created previously when running the conver-sites function.
}
\keyword{MQutil.mapSitesBack()}
\keyword{MSStats}
\keyword{map}
\keyword{mapping}
\keyword{ptm}
\keyword{results}
\keyword{sites}
