% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.MaxQToSaint}
\alias{MQutil.MaxQToSaint}
\title{Convert MaxQuant file into a format to work with SAINT}
\usage{
MQutil.MaxQToSaint(data_file, keys_file, ref_proteome_file,
  quant_variable = "spectral_count")
}
\arguments{
\item{keys_file}{The filepath to the keys file used in the MSStats analysis.}

\item{ref_proteome_file}{The filepath to the reference proteom (txt tab delimited file).}

\item{quant_variable}{Pick which quantitative value to use: "spectral_counts" or "ms1"}

\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}
}
\description{
Converts the MaxQuant evidence file to the 3 required files for SAINTexpress. One can choose to either use the spectral counts or the intensities for the analysis.
}
\keyword{MQutil.MaxQToSaint()}
\keyword{SAINT}
\keyword{intensity}
\keyword{ms1}
\keyword{spectral}
\keyword{spectral_counts}
