% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.MISTformat}
\alias{MQutil.MISTformat}
\title{Convert MaxQuant to MIST format}
\usage{
MQutil.MISTformat(input_file, keys_file, output_file, species = "HUMAN",
  uniprot_dir = "~/Box Sync/db/mist/")
}
\arguments{
\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}

\item{keys_file}{The filepath to the keys file used in the MSStats analysis.}

\item{output_file}{The filepath to the intended output file (txt tab delimited file).}

\item{species}{The species to in the data to be used to get weights from. Multiple species can be listed, separating with a '-'. (eg. HUMAN-MOUSE)}

\item{uniprot_dir}{The directory where the uniprot files are for the different species.}
}
\description{
Converts MaxQuant evidence file into a file format compatible with the MiST pipeline using MS/MS.Count. Note that this is the MiST data file, and that an additional keys file will have to be constructed before running MiST. Multiple species can be searched at once, simply separate them by a '-'. (eg. HUMAN-MOUSE).
}
\keyword{MIST}
\keyword{MQutil.MISTformat()}
\keyword{mist}
\keyword{spectral}
