% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.ProteinToSiteConversion}
\alias{MQutil.ProteinToSiteConversion}
\title{Convert Protein format to Site format}
\usage{
MQutil.ProteinToSiteConversion(maxq_file, ref_proteome_file, output_file,
  mod_type = "ub")
}
\arguments{
\item{maxq_file}{The filepath to the MaxQuant searched results file (txt tab delimited file).}

\item{ref_proteome_file}{Filepath to the fasta file containing the reference proteom.}

\item{mod_type}{The kind of PTM site analysis to be perfomed. Pick between ub (default), ph, and ac.}

\item{output}{The filepath to the intended output file (txt tab delimited file).}
}
\description{
This function is a preprocessing function that is used with PTM data. If you want to run a site specific analysis, before running MSStats on a UB/PH/AC set, you need to convert the evidence file into a format that MSStats will be able to diffentiate the sites with. This outputs a new file that should be used as the input file for your MSStats group conparison analysis.
}
\keyword{MQutil.ProteinToSiteConversion()}
\keyword{ac,}
\keyword{acetylation}
\keyword{ph,}
\keyword{phosphorylation,}
\keyword{site,}
\keyword{ub,}
\keyword{ubiquitination,}
