% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.sampleQuant}
\alias{MQutil.sampleQuant}
\title{Add Abundancd data to Results.}
\usage{
MQutil.sampleQuant(sq_file, contrast_file, results_file)
}
\arguments{
\item{sq_file}{The filepath to the Sample Quantification file (txt tab delimited file).}

\item{contrast_file}{The filepath to the Contrst file used to generate the MSstats results (txt tab delimited file).}

\item{results_file}{The filepath to the results in wide format file used (txt tab delimited file).}
}
\description{
Aggregates the normalized abundance and replicate data from the samples. Uses the MSstat output file ...mss-sampleQuant.txt for the aggregations, and is applied directly to the MSstats results in wide format. The resulting file will have 'abundance' appended to the end of the file name.
}
\keyword{MQutil.sampleQuant()}
\keyword{abundance}
\keyword{msstats}
\keyword{quatification}
\keyword{sample}
\keyword{samplequantification}
