% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R, R/MaxQ_utilities.R
\name{MQutil.SILACToLong}
\alias{MQutil.SILACToLong}
\alias{MQutil.SILACToLong}
\title{Convert the SILAC evidence file to MSstats format}
\usage{
MQutil.SILACToLong(filename, output)

MQutil.SILACToLong(filename, output)
}
\arguments{
\item{filename}{Text filepath to the evidence file}

\item{output}{Text filepath of the output name}

\item{filename}{The filepath to the MaxQuant silac data (txt tab delimited file).}

\item{output}{The filepath to the intended output file (txt tab delimited file).}
}
\value{
A data.frame with SILAC data processed for MSstats
}
\description{
Converting the evidence file from a SILAC search to a format
compatible with MSstats. It basically modifies the Raw.files adding the
Heavy and Light label

This function converts a silac labeled MaxQuant file and splits the H/L cases apart into separate rows.
}
\keyword{MQutil.SILACToLong()}
\keyword{silac}
