% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.spectralCounts}
\alias{MQutil.spectralCounts}
\title{Get Spectral Counts}
\usage{
MQutil.spectralCounts(input_file, keys_file, output_file)
}
\arguments{
\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}

\item{keys_file}{The filepath to the keys file used in the MSStats analysis.}

\item{output_file}{The filepath to the intended output file (txt tab delimited file).}
}
\description{
Outputs the spectral counts from the MaxQuant evidence file.
}
\keyword{MQutil.spectralCounts()}
\keyword{spectral}
\keyword{spectral_counts}
