% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msstat_summary.R
\name{msstats_summary}
\alias{msstats_summary}
\title{Summarize the MSStats results and data quantification}
\usage{
msstats_summary(evidence_file, prot_group_file, keys_file, results_file,
  return.results = FALSE)
}
\arguments{
\item{evidence_file}{The filepath to the MaxQuant searched data (evidence)
file (txt tab delimited file).}

\item{prot_group_file}{The filepath to the MaxQuant searched Protein Groups
file (txt tab delimited file).}

\item{keys_file}{The filepath to the keys file used with MSStats (txt tab
delimited file).}

\item{results_file}{The filepath to the MSStats results file in the default
long format (txt tab delimited file).}

\item{return.results}{Whether or not to return the results to the R
environment upon completion. This is useful if this is being used in an
R pipeline and you want to feed the results directly into the next stage
of analysis via an R environment/terminal. Regardless, the results will
be written to file. Default = FALSE}
}
\description{
Converts the MSStats results file to wide format
(columns are the contrasts), as well as adds BioReplicate information about
the Number of Unique Peptides, Spectral Counts, and Intensities for each
protein. In cases where there are multiple values for a Protein-BioReplicate
pair due to minute changes in sequence, the maximum value is taken for the
pair. Any pairs without a value are assigned a value of NA.
}
\keyword{MSStats}
\keyword{MaxQuant,}
\keyword{evidence,}
\keyword{msstats_summary()}
