% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_evidence}
\alias{ph_evidence}
\title{MaxQuant Phosphorylation evidence file}
\format{tab delimited file
\describe{
\item{ph_evidence}{The MaxQuant evidence file. Check this link to find out
more about it:
\url{http://www.coxdocs.org/doku.php?id=maxquant:start}}
}}
\usage{
ph_evidence
}
\description{
Comparison of the phosphoproteome of two head and neck cancer cell lines,
Cal33 and HSC6.
}
\details{
4 replicates per cell line

The information about the experimental design of the evidence file can
be found the data object called \code{ph_keys} also available in \code{artMS}
}
\keyword{datasets}
