% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_keys}
\alias{ph_keys}
\title{Keys file: Experimental design describing the \code{ph_evidence} data object}
\format{Tab delimited file with the following columns:
\describe{
\item{Raw.file}{Raw file processed. Each one should be either a
biological replicate or a technical replicate}

\item{IsotopeLabelType}{Type of labeling. \code{L} is used for label free
experiments}

\item{Condition}{Label for conditions. Only alpha-numeric characters and
\code{underscore (_)} are allowed}

\item{BioReplicate}{Label for the Biological replicates. The consensus is
to use the same label as the Condition, but adding a \code{dash (-)} corresponding
to the number of biological replicate. For example, \code{Cal-1}, \code{Cal-2},
\code{Cal-3}}
}}
\usage{
ph_keys
}
\description{
Data frame with the information about the experimental design.
It provides the metadata of the \code{ph_evidence} data object also available in
this package (see \code{vignettes} to find out more)
}
\details{
Comparison of the phosphoproteome of two head and neck cancer cell lines,
Cal33 and HSC6.

4 replicates per cell line
}
\keyword{datasets}
