% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{plotHeat}
\alias{plotHeat}
\title{Heatmap of significant values}
\usage{
plotHeat(mss_F, out_file, labelOrder = NULL, names = "Protein",
  cluster_cols = F, display = "log2FC")
}
\arguments{
\item{mss_F}{data.frame with the significant values (log2fc, pvalues)}

\item{out_file}{Name for the output}

\item{labelOrder}{Vector with the particular order for the IDs (default,
\code{NULL} no order)}

\item{names}{Type of ID used. Default is `Protein`` (uniprot entry id).
Soon will be possible to use 'Gene' name ids.}

\item{cluster_cols}{Select whether to cluster the columns. Options: \code{T}
or \code{F}. Default \code{T}.}

\item{display}{Value used to genarate the heatmaps. Options: \code{log2FC},
\code{adj.pvalue}, \code{pvalue}. Default: \code{log2FC}}
}
\value{
A heatmap of significant values
}
\description{
heatmap plot to represent proteins with significant changes
}
\keyword{heatmap}
\keyword{plotHeat()}
\keyword{significant,}
