% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{prettyPrintHeatmapLabels}
\alias{prettyPrintHeatmapLabels}
\title{Pretty Labels for Heatmaps}
\usage{
prettyPrintHeatmapLabels(uniprot_acs, uniprot_ids, gene_names)
}
\arguments{
\item{uniprot_acs}{Uniprot accession id}

\item{uniprot_ids}{Uniprot entry id}

\item{gene_names}{Gene symbol}
}
\value{
Pretty labels for a heatmap
}
\description{
Generates pretty labels for the heatmaps.
}
\keyword{plots,}
\keyword{pretty}
\keyword{prettyPrintHeatmapLabels()}
