% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomDF}
\alias{randomDF}
\title{Random data set}
\format{A data frame with 100 rows and 10 variables:
\describe{
Dataset generated using this code

\code{data.frame(replicate(10,sample(0:1,100,rep=TRUE)))}
}}
\usage{
randomDF
}
\description{
Dataset randomly generated for testing purposes
}
\keyword{datasets}
