% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evidence_file.R
\name{read_evidence_file}
\alias{read_evidence_file}
\title{Read in Evidence File}
\usage{
read_evidence_file(evidence_file)
}
\arguments{
\item{evidence_file}{The filepath to the MaxQuant searched data (evidence) file (txt tab delimited file).}
}
\description{
Read in a MaxQuant searched Evidence file using data.table. This function propperly classes each column and so fread doesn't have to guess.
}
\keyword{MaxQuant,}
\keyword{evidence}
\keyword{read_evidence_file()}
