% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToSaintExpressFormat.R
\name{removeMaxQProteinGroups}
\alias{removeMaxQProteinGroups}
\title{Remove protein groups}
\usage{
removeMaxQProteinGroups(data)
}
\arguments{
\item{data}{Data.frame with a \code{Proteins} column.}
}
\value{
A data.frame with the protein groups removed
}
\description{
Remove the group of proteins ids separated by separated by \code{;}
}
\keyword{maxquant,}
\keyword{proteingroups}
\keyword{remove,}
\keyword{removeMaxQProteinGroups()}
