% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsHeatmap.R,
%   R/resultsHeatmap_Interactive.R
\name{resultsHeatmap}
\alias{resultsHeatmap}
\alias{resultsHeatmap}
\title{Create Heatmap from MSStat Results}
\usage{
resultsHeatmap(results_file, pval = 0.05, prot_file = FALSE)

resultsHeatmap(results_file, pval = 0.05, prot_file = FALSE)
}
\arguments{
\item{results_file}{The filepath to the MSstats results file in the original long format.}

\item{results_file}{The filepath to the MSstats results file in the original long format.}
}
\description{
This will read in the MSStats results file (original long form) and output a heatmap that the user can interact with by hovering the mouse over it.

This will read in the MSStats results file (original long form) and output a heatmap that the user can interact with by hovering the mouse over it.
}
\keyword{MSStats}
\keyword{resultsHeatmap()}
