% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evidence_file.R
\name{unifyColumnNames}
\alias{unifyColumnNames}
\title{Unify column names}
\usage{
unifyColumnNames(x)
}
\arguments{
\item{x}{Loaded evidence file.}
}
\value{
A data.frame with the evidence file with defining classes and
unified columns
}
\description{
The newer versions of MaxQuant change the letter case in the
column names, which throws off the legacy RMSQ functions.
This is to help keep things unified still.
}
\keyword{MaxQuant,}
\keyword{evidence}
\keyword{unifyColumnNames()}
