% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{writeContrast}
\alias{writeContrast}
\title{Generate the contrast matrix required by MSstats from a txt file}
\usage{
writeContrast(contrast_file)
}
\arguments{
\item{contrast_file}{The text filepath of contrasts}
}
\description{
It simplifies the process of creating the contrast file
}
\keyword{writeContrast()}
