% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_analysisQuantifications}
\alias{artms_analysisQuantifications}
\title{Analysis of the Relative Quantification}
\usage{
artms_analysisQuantifications(log2fc_file, modelqc_file, specie,
  output_dir, enrich = TRUE, l2fc_thres = 1.5,
  choosePvalue = "adjpvalue", isBackground = "nobackground",
  isPtm = "global", mnbr = 2, isFluomics = FALSE,
  pathogen = "nopathogen")
}
\arguments{
\item{log2fc_file}{(char) MSstats results file location}

\item{modelqc_file}{(char) MSstats modelqc file location}

\item{specie}{(char) Main specie. Currently supported: human, mouse}

\item{output_dir}{(char) Name for the folder to output the results from the
function}

\item{enrich}{(logical) Performed enrichment analysis using GprofileR?
\code{TRUE} (default) or \code{FALSE}}

\item{l2fc_thres}{(int) log2fc cutoff for enrichment analysis (default,
\code{l2fc_thres = 1.5})}

\item{choosePvalue}{(char) specify whether \code{pvalue} or \code{adjpvalue} should
use for the analysis. The default option is \code{adjpvalue}
(multiple testing correction).
But if the number of biological replicates for a given experiment is
too low (for example n = 2), then \code{choosePvalue = pvalue} is recommended.}

\item{isBackground}{(char) background of gene names for enrichment analysis.
\code{nobackground} (default) will use the total number of genes detected.
Alternatively provided the file path name to the background gene list.}

\item{isPtm}{(char) Is a ptm-site quantification?
\itemize{
\item \code{global} (default),
\item \code{ptmsites} (for site specific analysis),
\item \code{ptmph} (Jeff's script output evidence file)
}}

\item{mnbr}{(int) minimal number of biological replicates for imputation
and filtering. Default: \code{mnbr = 2} (Proteins must be found in one of the
conditions in at least 2 of the biological replicates)}

\item{isFluomics}{(logical) Does this data belong to the FluOMICs project?
\code{TRUE} or \code{FALSE} (default)}

\item{pathogen}{(char) Is there a pathogen in the dataset as well?
if it does not, then use \code{pathogen = nopathogen} (default).
Pathogens available: \code{tb} (Tuberculosis), \code{lpn} (Legionella)}
}
\value{
(data.frame) summary of quantifications, including annotations,
enrichments, etc
}
\description{
Analysis of the Relative Quantifications obtained by MSstats.
It includes:
\itemize{
\item Annotations
\item Summary files in different format (xls, txt) and shapes (long, wide)
\item Numerous summary plots
\item Enrichment analysis using Gprofiler
\item PCA of protein abundance
\item PCA of quantifications
\item Clustering analysis
}
}
\examples{
# Testing that the files cannot be empty
artms_analysisQuantifications(log2fc_file = NULL,
                              modelqc_file = NULL,
                              specie = NULL,
                              output_dir = NULL)
}
\keyword{analysis,}
\keyword{quantifications}
