% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisQuantifications.R
\name{artms_analysisQuantifications}
\alias{artms_analysisQuantifications}
\title{Analysis of the Relative Quantification}
\usage{
artms_analysisQuantifications(log2fc_file, modelqc_file, specie,
  enrich = TRUE, output_dir, isFluomics = FALSE, isPtm = "noptm",
  isBackground, mnbr, threshold, ipval, pathogen = "nopathogen")
}
\arguments{
\item{log2fc_file}{MSstats results file}

\item{modelqc_file}{MSstats modelqc file}

\item{specie}{Specie (human, mouse)}

\item{enrich}{Performed enrichment analysis? \code{TRUE} (default) or \code{FALSE}}

\item{output_dir}{results folder name}

\item{isFluomics}{Is from the fluomics project? \code{TRUE} or \code{FALSE} (default)}

\item{isPtm}{Is a ptm quantification? noptm (default), yesptmph, yesptmsites}

\item{isBackground}{background gene set}

\item{mnbr}{minimal number of biological replicates for imputation}

\item{threshold}{log2fc cutoff for enrichment analysis}

\item{ipval}{pvalue cutoff for enrichment analysis}

\item{pathogen}{is there a pathogen in the dataset as well? if it does not,
then use \code{pathogen = nopathogen} (default), \code{tb} (Tuberculosis),
\code{lpn} (Legionella)}
}
\value{
summary of quantifications, including annotations, enrichments, etc
}
\description{
Analysis of the Relative Quantifications obtained by MSstats.
It includes:
\itemize{
\item Annotations
}
}
\keyword{analysis,}
\keyword{artms_analysisQuantifications()}
\keyword{quantifications}
