% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{artms_annotationUniprot}
\alias{artms_annotationUniprot}
\title{Annotate table with Gene Symbol and Name based on Uniprot ID(s)}
\usage{
artms_annotationUniprot(data, columnid, sps)
}
\arguments{
\item{data}{(data.frame) to be annotated (or file path and name)}

\item{columnid}{(char) The column with the uniprotkb ids}

\item{sps}{(char) The specie name. Check \code{?artms_mapUniprot2entrezGeneName}
to find out more about supported species.}
}
\value{
(data.frame) with two new columns: \code{Gene} and \code{Protein.name}
}
\description{
Annotate gene name and symbol based on uniprot ids. It will take
the column from your data.frame specified by the \code{columnid} argument,
search for the gene symbol, name, and entrez based on the specie (\code{sps}
argument) and merge the information back to the input data.frame
}
\examples{
# This example adds annotations to the evidence file available in 
# artMS, based on the column "Proteins".

evidence_anno <- artms_annotationUniprot(data = artms_data_ph_evidence,
                                         columnid = "Proteins", 
                                         sps = "human")
}
\keyword{annotation,}
\keyword{uniprot}
