% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_intensity_RT.R
\name{artms_avg_intensity_RT}
\alias{artms_avg_intensity_RT}
\title{Summarize average intensity and retention time per protein}
\usage{
artms_avg_intensity_RT(evidence_file, protein_file = NULL,
  output_file = FALSE, species, verbose = TRUE)
}
\arguments{
\item{evidence_file}{(char) The filepath to the MaxQuant searched data
(evidence) file (txt tab delimited file).}

\item{protein_file}{(char) The filepath to a file or vector conatining
a list of proteins of interest.}

\item{output_file}{(char) The file name for the results
(must have the extension \code{.txt}). If empty, then the
results will be returned as an R object.}

\item{species}{(char) The species name.
Check \code{?artmsMapUniprot2Entrez} for supported species}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
An R object with the results and a file with the results (if the
output_file argument is provided). It contains averages of Intensity,
Retention Time, Caliberated Retention Time
}
\description{
Input an evidence file from MaxQuant and a file
containing a list of proteins of interest (optional).
The function will summarize from the evidence file and report back the
average intensity, average retention time, and the average caliberated
retention time. If a list of proteins is provided, then only those proteins
will be summarized and returned.
}
\examples{
ave_int <- artms_avg_intensity_RT(evidence_file = artms_data_ph_evidence,
                                  species = "human")
}
\keyword{MaxQuant,}
\keyword{caliberated}
\keyword{evidence,}
\keyword{intensity,}
\keyword{retention}
\keyword{summary,}
\keyword{time,}
