% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_evidence}
\alias{artms_data_ph_evidence}
\title{Evidence file example}
\format{A data frame with all the columns available in an evidence file
generated with MaxQuant version 1.6.2.3}
\usage{
artms_data_ph_evidence
}
\description{
Evidence file from a PH experiment consisting of two
head and neck cancer cell lines ("Conditions" \code{"Cal33"} and \code{"HSC6"}).

The number of peptides was reduced to 1/5 due
to bioconductor limitations on data size.
}
\keyword{datasets}
