% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_evidence}
\alias{artms_data_ph_evidence}
\title{Evidence file example}
\format{
A data frame with all the columns available in an evidence file
generated with MaxQuant version 1.6.2.3
}
\usage{
artms_data_ph_evidence
}
\description{
Evidence file from a PH experiment consisting of two
head and neck cancer cell lines ("Conditions" \code{"Cal33"} and \code{"HSC6"}).

Unfortunately, the number of lines was reduced to 1/8 due
to bioconductor limitations on data size, which means that this data is
not very representative of a real evidence file. However, both the
full evidence.txt and keys.txt file are available at:
http://kroganlab.ucsf.edu/artms/ph/evidence.txt
http://kroganlab.ucsf.edu/artms/ph/keys.txt
}
\keyword{datasets}
