% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artms_data_ph_evidence}
\alias{artms_data_ph_evidence}
\title{Evidence file example}
\format{A data frame with all the columns available in an evidence file
generated with MaxQuant version 1.6.2.3}
\usage{
artms_data_ph_evidence
}
\description{
Reduced version of a Evidence file from a PH experiment. Only
two biological replicates are available, and the number of peptides was
reduced to 1/5 due to bioconductor limitations on data size.
}
\keyword{datasets}
