% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichLog2fc}
\alias{artms_enrichLog2fc}
\title{Enrichment of changes in protein abundance or PTMs}
\usage{
artms_enrichLog2fc(data, fileout, specie, background)
}
\arguments{
\item{data}{data.frame with \code{Gene} and \code{Comparison} columns}

\item{fileout}{Name for the output file}

\item{specie}{Specie, only supported "human" or "mouse"}

\item{background}{Background genes for the enrichment analysis.}
}
\value{
Results from the enrichment analysis using Gprofiler
}
\description{
Enrichment analysis of the selected proteins
}
\keyword{artms_enrichLog2fc()}
\keyword{enrichment}
