% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichLog2fc}
\alias{artms_enrichLog2fc}
\title{Enrichment of changes in protein abundance or PTMs}
\usage{
artms_enrichLog2fc(dataset, species, background, heatmaps = FALSE,
  output_name = "enrichment.txt", verbose = TRUE)
}
\arguments{
\item{dataset}{(data.frame) with a \code{Gene} and \code{Comparison or Label} (with
the name of the comparisons specified in the contrast file) columns}

\item{species}{(char) Specie, only supported "human" or "mouse"}

\item{background}{(vector) Background genes for the enrichment analysis.}

\item{heatmaps}{(logical) if \code{TRUE} generates heatmaps (pdf),
\code{FALSE} (default) otherwise.}

\item{output_name}{(char) Name of the annotation files, which will be used
as well for the heatmaps (if \code{heatmaps} is selected)
Default \code{output_name = "enrichment.txt"}}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
(data.frame) Results from the enrichment analysis using Gprofiler
and heatmaps (if selected)
}
\description{
Enrichment analysis of the selected proteins
}
\examples{
# The data must be annotated (Protein and Gene columns)
data_annotated <- artms_annotationUniprot(
                      x = artms_data_ph_msstats_results,
                      columnid = "Protein",
                      species = "human")
# And then the enrichment
enrich_set <- artms_enrichLog2fc(
                   dataset = data_annotated,
                   species = "human",
                   background = unique(data_annotated$Gene))
}
\keyword{enrichment}
