% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{artms_enrichProfiler}
\alias{artms_enrichProfiler}
\title{Enrichment analysis using GprofileR}
\usage{
artms_enrichProfiler(x, categorySource = c("GO"), specie,
  background = NA)
}
\arguments{
\item{x}{List of protein ids. \code{x} can be anything: either a list of ids,
or you could also send a data.frame and it will find the columns
with the IDs. Is not cool?}

\item{categorySource}{Resources providing the terms on which the enrichment
will be performed. The supported resources by gprofiler are:
\itemize{
\item \code{GO (GO:BP, GO:MF, GO:CC)}: Gene Ontology (see more below)
\item \code{KEGG}: Biological pathways
\item \code{REAC}: Biological pathways (Reactome)
\item \code{TF}: Regulatory motifs in DNA (TRANSFAC TFBS)
\item \code{MI}: Regulatory motifs in DNA (miRBase microRNAs)
\item \code{CORUM}: protein complexes database
\item \code{HP}: Human Phenotype Ontology
\item \code{HPA}: Protein databases (Human Protein Atlas)
\item \code{OMIM}: Online Mendelian Inheritance in Man annotations:
\item \code{BIOGRID}: BioGRID protein-protein interactions
The type of annotations for Gene Ontology:
\item Inferred from experiment \code{\link{IDA, IPI, IMP, IGI, IEP}}
\item Direct assay \code{\link{IDA}} / Mutant phenotype \code{\link{IMP}}
\item Genetic interaction \code{\link{IGI}} / Physical interaction \code{\link{IPI}}
\item Traceable author \code{\link{TAS}} / Non-traceable author \code{\link{NAS}} /
Inferred by curator \code{\link{IC}}
\item Expression pattern \code{\link{IEP}} / Sequence or structural similarity \code{\link{ISS}}
/ Genomic context \code{\link{IGC}}
\item Biological aspect of ancestor \code{\link{IBA}} / Rapid divergence \code{\link{IRD}}
\item Reviewed computational analysis \code{\link{RCA}} / Electronic annotation \code{\link{IEA}}
\item No biological data \code{\link{ND}} / Not annotated or not in background \code{\link{NA}}
}}

\item{specie}{Specie code: Organism names are constructed by concatenating
the first letter of the name and the family name.
Example: human - ’hsapiens’, mouse - ’mmusculus’.}

\item{background}{Vector gene list to use as background for the enrichment
analysis. Default: \code{NA}}
}
\value{
The enrichment results as provided by gprofiler
}
\description{
This function simplifies the enrichment analysis performed by
the excellent tool GprofileR.
}
\details{
This function uses the following \code{gprofiler} arguments as default:
\itemize{
\item ordered_query = F
\item significant = T
\item exclude_iea = T
\item underrep = F
\item evcodes = F
\item region_query = F
\item max_p_value = 0.05
\item min_set_size = 0
\item max_set_size = 0
\item min_isect_size = 0
\item correction_method = "analytical" #Options: "gSCS", "fdr", "bonferroni"
\item hier_filtering = "none"
\item domain_size = "known" # annotated or known
\item numeric_ns = ""
\item png_fn = NULL
\item include_graph = T
}
}
\keyword{artms_enrichProfiler()}
\keyword{enrichment}
