% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceQC.R
\name{artms_evidenceQC}
\alias{artms_evidenceQC}
\title{Quality Control analysis of the MaxQuant evidence file}
\usage{
artms_evidenceQC(evidence_file, keys_file, prot_exp, fractions = 0)
}
\arguments{
\item{evidence_file}{(char) The evidence file path and name}

\item{keys_file}{(char) The keys file path and name}

\item{prot_exp}{(char) Proteomics experiment. 4 options available:
\itemize{
\item \code{APMS}: affinity purification mass spectrometry
\item \code{AB}: protein abundance
\item \code{PH}: protein phosphorylation
\item \code{UB}: protein ubiquitination (aka ubiquitylation)
}}

\item{fractions}{(binary) Is a fractionated experiment?
\itemize{
\item 1 yes
\item 0 no (default)
}}
}
\value{
Quality control files and plots
}
\description{
Quality Control analysis of the MaxQuant evidence file
}
\examples{
\donttest{
artms_evidenceQC(evidence_file = "evidence.txt", 
                 keys_file = "keys.txt", 
                 prot_exp = "PH")
}
}
\keyword{QC,}
\keyword{control,}
\keyword{evidence}
\keyword{quality,}
