% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToSaintExpressFormat.R
\name{artms_evidenceToSaintExpressFormat}
\alias{artms_evidenceToSaintExpressFormat}
\title{MaxQuant evidence file to SAINTexpress format}
\usage{
artms_evidenceToSaintExpressFormat(data_file, keys_file, ref_proteome_file,
  quant_variable = "spectral_count", output_file)
}
\arguments{
\item{data_file}{Maxquant evidence file}

\item{keys_file}{Keys file with a SAINT column specifying the test (\code{T}) and
control (\code{C}) conditions}

\item{ref_proteome_file}{Reference proteome file in fasta format}

\item{quant_variable}{choose either \code{spectral_counts} (spectral counts)
or \code{ms1}}

\item{output_file}{Output file name}
}
\value{
The 3 required files by SAINTexpress
}
\description{
Converts the MaxQuant evidence file to the 3 required files
for SAINTexpress. One can choose to either use the \code{spectral counts}
(use \code{spectral_count}) or the \code{intensities} (use \code{ms1}) for the analysis.
}
\keyword{APMS}
\keyword{SAINT,}
\keyword{SAINTexpress,}
\keyword{artms_EvidenceToSaintExpressFormat()}
