% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_filterMaxqData}
\alias{artms_filterMaxqData}
\title{Remove contaminants and empty proteins}
\usage{
artms_filterMaxqData(data)
}
\arguments{
\item{data}{the data.frame in long format}
}
\value{
A new data.frame without REV__ and CON__ Protein ids
}
\description{
Remove contaminants and erronously identified 'reverse'
sequences by MaxQuant
}
\keyword{artms_filterMaxqData()}
\keyword{cleanup,}
\keyword{contaminants}
