% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfFile.R
\name{artms_isEvidenceNewVersion}
\alias{artms_isEvidenceNewVersion}
\title{Check if a given evidencee file was generated by a new version of
MaxQuant (v>1)}
\usage{
artms_isEvidenceNewVersion(evidence_file)
}
\arguments{
\item{evidence_file}{the evidence file name}
}
\value{
(logical) \code{TRUE} if it is a newer version of MaxQuant,
\code{FALSE} otherwise
}
\description{
MaxQuant introduced changes in the column names and number
of columns for the evidence file in version 1 (we think).
This function check whether the evidence comes from the latest version
of MaxQuant.
}
\examples{
artms_isEvidenceNewVersion(evidence_file = artms_data_ph_evidence)
}
\keyword{check}
\keyword{evidence,}
\keyword{file,}
\keyword{input,}
\keyword{version}
