% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_mergeEvidenceKeysByFiles}
\alias{artms_mergeEvidenceKeysByFiles}
\title{Merge evidence and keys by file name}
\usage{
artms_mergeEvidenceKeysByFiles(evidence_file, keys_file)
}
\arguments{
\item{evidence_file}{(char) The Evidence file name}

\item{keys_file}{(char) The keys file name}
}
\value{
(data.frame) with both evidence and keys files merged by raw.files
}
\description{
Merge evidence and keys by file name
}
\examples{
\donttest{
evidenceKeys <- artms_mergeEvidenceKeysByFiles(
	evidence_file = "FLU-THP1-H1N1-AB-evidence.txt", 
	keys_file = "FLU-THP1-H1N1-AB-keys.txt")
}
}
\keyword{evidence,}
\keyword{internal,}
\keyword{keys}
\keyword{merge,}
