% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_mergeMaxQDataWithKeys}
\alias{artms_mergeMaxQDataWithKeys}
\title{Merge evidence and keys files}
\usage{
artms_mergeMaxQDataWithKeys(data, keys, by = c("RawFile"))
}
\arguments{
\item{data}{The evidence in data.frame}

\item{keys}{The keys in data.frame}

\item{by}{(vector) specifying the columns use to merge the evidence and keys.
Obviously, both data.frames must have this column name.}
}
\value{
(data.frame) with the evidence and keys merged
Default column to merge: \code{RawFile}
}
\description{
Merge the evidence and keys files on the given columns
}
\examples{
\donttest{
evidence <- read.delim("FLU-THP1-H1N1-AB-evidence.txt", stringsAsFactors = F)
keys <- read.delim("FLU-THP1-H1N1-AB-keys.txt", stringsAsFactors = F)
evidenceKeys <- artms_mergeMaxQDataWithKeys(data = evidence, keys = keys)
}
}
\keyword{evidence,}
\keyword{keys}
\keyword{merge,}
