% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{artms_plotReproducibilityEvidence}
\alias{artms_plotReproducibilityEvidence}
\title{Generate reproducibility plots based on raw intentities
(from the evidence file)}
\usage{
artms_plotReproducibilityEvidence(data)
}
\arguments{
\item{data}{evidence data.frame}
}
\value{
A reproducibility plot based on evidence values
}
\description{
Generate reproducibility plots based on raw intentities
(log tranformed) from the evidence file
}
\keyword{artms_plotReproducibilityEvidence()}
\keyword{control}
\keyword{plot,}
\keyword{qc,}
\keyword{quality,}
