% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinToSiteConversion.R
\name{artms_proteinToSiteConversion}
\alias{artms_proteinToSiteConversion}
\title{Converts the \code{Proteins} column of the evidence file to site-specific
\code{Uniprot_PTM} notation}
\usage{
artms_proteinToSiteConversion(evidence_file, ref_proteome_file,
  output_file, mod_type = "PH")
}
\arguments{
\item{evidence_file}{(char) The evidence file name and location}

\item{ref_proteome_file}{(char) The reference proteome used as database
to search the \code{evidence.txt} file with MaxQuant. It will be used to map the
modified peptide to the protein sequence and find the site location.
Therefore, it does NOT need the MaxQuant's \code{Phospho (STY)Sites.txt}}

\item{output_file}{(char) Output file name (\code{-sites-evidence.txt} recommended)}

\item{mod_type}{(char) The posttranslational modification. Options:
\itemize{
\item \code{ub}: Protein Ubiquitination
\item \code{ph}: Protein Phosphorylation
\item \code{ac}: Protein Acetylation
}}
}
\value{
(file) Return a new evidence file with the \code{Proteins} column
modified by adding the sequence site location(s) + postranslational
modification(s) to the uniprot entry id.
Examples: \code{A34890_ph3}; \code{Q64890_ph24_ph456}; \code{Q64890_ub34_ub129_ub234};
\code{Q64890_ac35}.
}
\description{
It enables the site-specific quantification of PTMs by
converting the \code{Proteins} column of the evidence file to an \code{Uniprot_PTM}
notation. In this way, each of the modified peptides can be quantified
independently across conditions
}
\examples{
\donttest{
artms_proteinToSiteConversion(
  evidence_file = "evidence.txt", 
  ref_proteome_file = "uniprot_canonical.fasta", 
  output_file = "phsite_evidence.txt", 
  mod_type = "ph")
}
}
\keyword{ac}
\keyword{convert,}
\keyword{evidence,}
\keyword{ph,}
\keyword{ptm,}
\keyword{ub,}
