% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinToSiteConversion.R
\name{artms_proteinToSiteConversion}
\alias{artms_proteinToSiteConversion}
\title{Converts the \code{Proteins} column of the evidence file to site-specific
\code{Uniprot_PTM} notation}
\usage{
artms_proteinToSiteConversion(maxq_file, ref_proteome_file, output_file,
  mod_type = "ub")
}
\arguments{
\item{maxq_file}{The evidence file name and location}

\item{ref_proteome_file}{The reference proteome (to map the peptide
to the protein sequence and find out the site location)}

\item{output_file}{Output file name (-sites-evidence.txt recommended)}

\item{mod_type}{The posttranslational modification. Options: \code{ub}, \code{ph}, or
\code{ac}}
}
\value{
Return a new evidence file with the \code{Proteins} column modified by
adding the sequence site location(s) + postranslational modification(s)
to the uniprot entry id.
Examples: \code{A34890_ph3}; \code{Q64890_ph24_ph456};
\code{Q64890_ub34_ub129_ub234}; \code{Q64890_ac35}.
}
\description{
It enables the site-specific quantification of PTMs by
converting the \code{Proteins} column of the evidence file to a \code{Uniprot_PTM}
notation. In this way, each of the modified peptides can be quantified
independently.
}
\keyword{artms_proteinToSiteConversion()}
