% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{artms_quantification}
\alias{artms_quantification}
\title{Relative quantification using MSstats}
\usage{
artms_quantification(yaml_config_file)
}
\arguments{
\item{yaml_config_file}{(char) The yaml file name and location}
}
\value{
All the selected options
}
\description{
Relative quantification using MSstats
}
\keyword{artms_quantification()}
\keyword{driver,}
\keyword{function}
\keyword{main,}
