% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicatePlots.R
\name{artms_replicatePlots}
\alias{artms_replicatePlots}
\title{Create the replicate plots based on the pairings from
the replicate plot file}
\usage{
artms_replicatePlots(input_file, keys_file, replicate_file, out_file)
}
\arguments{
\item{input_file}{MaxQuant evidence file and location}

\item{keys_file}{Keys file with the experimental details}

\item{replicate_file}{Replicate file}

\item{out_file}{Output text file with the intensity values of every feature}
}
\value{
The output file of the summary of features and intensity values
}
\description{
Outputs a replicate plots based on a user provied file
containing the replicates to be compared. Values are based on the log2
value of the maximum intensities per modified sequence.
The "replicate file" should describe which replicates of which conditions
should be compared against each other. Each row represents a replicate
plot to be created. The file should be structured using the following
format and column names:

| condition1 | rep1_1 | rep1_2 | condition2 | rep2_1 | rep2_2 |
|---|---|---|---|---|---|
|  Infected | Infected_Rep1_name | Infected_Rep2_name | Negative | Negative_Rep1_name | Negative_Rep2_name |
| etc... |   |   |   |   |   |
| etc... |   |   |   |   |   |
}
\keyword{artms_replicatePlots()}
\keyword{evidence,}
\keyword{plots}
\keyword{replica,}
