% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{artms_SILACtoLong}
\alias{artms_SILACtoLong}
\title{Convert the SILAC evidence file to MSstats format}
\usage{
artms_SILACtoLong(evidence_file, output)
}
\arguments{
\item{output}{(char) Text filepath of the output name}

\item{filename}{(char) Text filepath to the evidence file}
}
\value{
(data.frame) with SILAC data processed for MSstats (and output file)
}
\description{
Converting the evidence file from a SILAC search to a format
compatible with MSstats. It basically modifies the Raw.files adding the
Heavy and Light label
}
\examples{
\donttest{
evidence2silac <- artms_SILACtoLong(evidence_file = "silac.evicence.txt", 
                                   output = "silac-evidence.txt")
}
}
\keyword{convert,}
\keyword{evidence}
\keyword{silac,}
