% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{artms_volcanoPlot}
\alias{artms_volcanoPlot}
\title{Volcano plot (log2fc / pvalues)}
\usage{
artms_volcanoPlot(mss_results_sel, lfc_upper, lfc_lower, FDR, file_name = "",
  PDF = T, decimal_threshold = 16)
}
\arguments{
\item{mss_results_sel}{Selected MSstats results}

\item{lfc_upper}{log2fc upper threshold (positive value)}

\item{lfc_lower}{log2fc lower threshold (negative value)}

\item{FDR}{False Discovery Rate threshold}

\item{file_name}{Name for the output file}

\item{PDF}{Option to generate pdf format. Default: \code{T}}

\item{decimal_threshold}{Decimal threshold for the pvalue.
Default: 16 (10^-16)}
}
\description{
It generates a scatter-plot used to quickly identify changes
}
\keyword{artms_volcanoPlot()}
\keyword{plot,}
\keyword{volcano}
