% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{artmsConvertMetabolomics}
\alias{artmsConvertMetabolomics}
\title{Convert Markview Metabolomics file (alignment table) into
a artMS compatible format}
\usage{
artmsConvertMetabolomics(input_file, out_file, id_file = NULL,
  verbose = TRUE)
}
\arguments{
\item{input_file}{(char) Markview input file}

\item{out_file}{(char) Output file name}

\item{id_file}{(char) KEGG database}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
(text file) Outputs the converted output name
}
\description{
MarkerView is an ABSciex software that supports the files
generated by Analyst software (.wiff) used to run our specific mass
spectrometer (ABSciex Triple TOF 5600+).
It also supports .t2d files generated by the
Applied Biosystems 4700/4800 MALDI-TOF.
Once the alignment is done by the software,
it can also run statistical analysis
(t-tests, unsupervised or supervised PCA).
Moreover, it attempts to categorize features as "Monoisotopic" vs "Isotopic"
if possible, allowing further filtering of the feature list.
The alignment table is automatically generated by Markerview and has a
unique file type (.mrkvw). The table can be copied and pasted into Excel
and then converted to a .txt file. Features identified as "isotopes"
are removed before the table is generated.
This file is further processed and formatted to perform QC and relative
quantification using the artMS functions
}
\examples{
\donttest{
artmsConvertMetabolomics(input_file = "markview-alignment.txt", 
                         out_file = "metabolomics-evidence.txt", 
                         id_file=NULL)
}
}
\keyword{convert}
\keyword{metabolomics,}
