% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherApps.R
\name{artmsPhosfateOutput}
\alias{artmsPhosfateOutput}
\title{Generate Phosfate Input file}
\usage{
artmsPhosfateOutput(inputFile)
}
\arguments{
\item{inputFile}{(char) the \code{imputedL2fcExtended.txt} file name and location}
}
\value{
Multiple output files (inputs of phosfate)
}
\description{
It takes as input the \code{imputedL2fcExtended.txt} results
generated by the \code{artms_analysisQuantifications()} function and generates
the \href{http://phosfate.com/}{Phosfate} input file
}
\examples{
\donttest{
artmsPhosfateOutput(inputFile)
}
}
\keyword{files}
\keyword{generate,}
\keyword{outputs,}
