% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherApps.R
\name{artmsPhotonOutput}
\alias{artmsPhotonOutput}
\title{Generate PHOTON Input file}
\usage{
artmsPhotonOutput(inputFile, output_dir = ".", verbose = TRUE)
}
\arguments{
\item{inputFile}{(char) the \code{imputedL2fcExtended.txt} file name and location}

\item{output_dir}{(char) Name of the folder to output results
(Default: current. Recommended: "photon_input_files" or similar)}

\item{verbose}{(logical) \code{TRUE} (default) to show function messages}
}
\value{
Multiple output files (inputs of phosfate)
}
\description{
It takes as input the \code{imputedL2fcExtended.txt} results
generated by the \code{artmsAnalysisQuantifications()} function and generates
the \href{https://github.com/jdrudolph/photon}{PHOTON} input file.
Please, notice that the only species suported by PHOTON is humans.
}
\examples{
\donttest{
artmsPhotonOutput(inputFile)
}
}
\keyword{files}
\keyword{generate,}
\keyword{outputs,}
