% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinToSiteConversion.R
\name{artmsProtein2SiteConversion}
\alias{artmsProtein2SiteConversion}
\title{Converts the \code{Proteins} column of the evidence file to site-specific
\code{Uniprot_PTM} notation}
\usage{
artmsProtein2SiteConversion(evidence_file, ref_proteome_file, output_file,
  mod_type, verbose = TRUE)
}
\arguments{
\item{evidence_file}{(char) The evidence file name and location}

\item{ref_proteome_file}{(char) The reference proteome used as database
to search the \code{evidence.txt} file with MaxQuant. It will be used to map the
modified peptide to the protein sequence and find the site location.
Therefore, it does NOT need the MaxQuant's \code{Phospho (STY)Sites.txt}}

\item{output_file}{(char) Output file name
(\code{-sites-evidence.txt} recommended)}

\item{mod_type}{(char) The posttranslational modification. Options:
\itemize{
\item \code{UB}: Protein Ubiquitination
\item \code{PH}: Protein Phosphorylation
\item \code{AC}: Protein Acetylation
}}

\item{verbose}{(logical) \code{TRUE} (default) shows function messages}
}
\value{
(file) Return a new evidence file with the \code{Proteins} column
modified by adding the sequence site location(s) + postranslational
modification(s) to the uniprot entry id.

Examples: \code{A34890_ph3}; \code{Q64890_ph24_ph456}; \code{Q64890_ub34_ub129_ub234};
\code{Q64890_ac35}.
}
\description{
It enables the site-specific quantification of PTMs by
converting the \code{Proteins} column of the evidence file to an \code{Uniprot_PTM}
notation. In this way, each of the modified peptides can be quantified
independently across conditions
}
\examples{
# Testing warning if files are not submitted
artmsProtein2SiteConversion(evidence_file = NULL, ref_proteome_file = NULL, 
output_file = NULL)
}
\keyword{ac}
\keyword{convert,}
\keyword{evidence,}
\keyword{ph,}
\keyword{ptm,}
\keyword{ub,}
