% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R
\name{mergeMaxQDataWithKeys}
\alias{mergeMaxQDataWithKeys}
\title{Merge MaxQuant Data with Keys}
\usage{
mergeMaxQDataWithKeys(data, keys, by = c("RawFile"))
}
\arguments{
\item{data}{Data from the evidence file in a data.table .}

\item{keys}{The information from the keys file in a data.table .}

\item{by}{What column(s) to merge the data and keys data on. Default = 'RawFile'}
}
\description{
This will merge a MaxQuant evidence file and keys file.
}
\keyword{MSStats}
\keyword{MaxQuant}
\keyword{mergeMaxQDataWithKeys()}
\keyword{mss}
