% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.resultsWide}
\alias{MQutil.resultsWide}
\title{Convert default MSStat results file to Wide format.}
\usage{
MQutil.resultsWide(input_file, save_file = F, return_results = F)
}
\arguments{
\item{input_file}{The filepath to the MaxQuant searched results file (txt tab delimited file). This can either be in either long or wide format.}

\item{save_file}{The filepath to the intended output file (txt tab delimited file).}

\item{return_results}{Whether to return the results at the end of the function (default =F).}
}
\description{
Converts the normal MSStats output file into 'wide' format where each row represents a protein's results, and each column represents the comparison made by MSStats. The fold change and p-value of each comparison will be it's own column.
}
\keyword{MQutil.resultsWide()}
\keyword{MSStats}
\keyword{results}
\keyword{wide}
