% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQ_utilities.R
\name{MQutil.SILACToLong}
\alias{MQutil.SILACToLong}
\title{Convert Silac to Long Format}
\usage{
MQutil.SILACToLong(filename, output)
}
\arguments{
\item{filename}{The filepath to the MaxQuant silac data (txt tab delimited file).}

\item{output}{The filepath to the intended output file (txt tab delimited file).}
}
\description{
This function converts a silac labeled MaxQuant file and splits the H/L cases apart into separate rows.
}
\keyword{MQutil.SILACToLong()}
\keyword{silac}
