% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_evidence}
\alias{ph_evidence}
\title{MaxQuant Phosphorylation evidence file}
\format{tab delimited file
\describe{
\item{ph_evidence}{The MaxQuant evidence file. Check this link to find out
more about it:
\url{http://www.coxdocs.org/doku.php?id=maxquant:start}}
}}
\usage{
ph_evidence
}
\description{
Phosphorylation data set generated on Human cell lines
}
\details{
Two different cell lines analyzed:  \code{Cal33} and \code{HSC6}

4 replicates per cell line

The information about the experimental design of the evidence file can
be found in the data.frame also available in this package named \code{ph_keys}
}
\keyword{datasets}
