% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsHeatmap.R
\name{resultsHeatmap}
\alias{resultsHeatmap}
\title{Create Heatmap from MSStat Results}
\usage{
resultsHeatmap(results_file, pval = 0.05, prot_file = FALSE,
  save_file = FALSE)
}
\arguments{
\item{results_file}{The filepath to the MSstats results file in the original long format.}

\item{prot_file}{A file listing proteins to filter out and analyze with the volcano plot (Default = F).}

\item{save_file}{A file path to where the heatmap pdf is to be saved.}
}
\description{
This will read in the MSStats results file (original long form) and output a heatmap of the results. The plot, as a ggplot object, is returned.
}
\keyword{MSStats}
\keyword{resultsHeatmap()}
