% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstats_functions.R, R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot (log2fc / pvalues)}
\usage{
volcanoPlot(mss_results_sel, lfc_upper = 1, lfc_lower = -1, FDR = 0.05,
  PDF = F, file_name = "volcano.pdf", decimal_threshold = 16)

volcanoPlot(mss_results_sel, lfc_upper = 1, lfc_lower = -1, FDR = 0.05,
  PDF = F, file_name = "volcano.pdf", decimal_threshold = 16)
}
\arguments{
\item{mss_results_sel}{Selected MSstats results}

\item{lfc_upper}{log2fc upper threshold (positive value)}

\item{lfc_lower}{log2fc lower threshold (negative value)}

\item{FDR}{False Discovery Rate threshold}

\item{PDF}{Option to generate pdf format. Default: \code{T}}

\item{file_name}{Name for the output file}

\item{decimal_threshold}{Decimal threshold for the pvalue.
Default: 16 (10^-16)}

\item{mss_results_sel}{The filepath to the MSstats results file (txt tab delimited file).}

\item{file_name}{The filepath and file name to the intended output file (e.g \code{volcano.pdf}).}

\item{lfc_upper}{The starting point of the upper region on interest based on the Log2FC (default = 1).}

\item{lfc_lower}{The starting point of the lower region on interest based on the Log2FC (default = -1).}

\item{FDR}{The adjusted p-value cutoff to define the region of interest (default = 0.05).}

\item{PDF}{Whether or not to write ou a pdf of the results or not. (default = F).}

\item{decimal_threshold}{Apply a decimal threshold. (default = 16).}
}
\description{
It generates a scatter-plot used to quickly identify changes

This function creates volcanon plots of the MSstats results, allowing the user to define regions of interest (confident hits) based on the Log2FC and adjusted p-value. A ggplot object will be returned.
}
\keyword{MSstats}
\keyword{plot,}
\keyword{volcano}
\keyword{volcano,}
\keyword{volcanoPlot()}
